% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ace.R
\docType{data}
\name{Ace}
\alias{Ace}
\title{Ace}
\format{\code{\link{R6Class}} object.}
\usage{
Ace
}
\description{
Ace
}
\section{Fields}{

\describe{
\item{\code{principals}}{list of class \code{\link{Principal}}.}

\item{\code{privileges}}{list of class \code{\link{Privilege}}.}
}}

\keyword{datasets}
