% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\docType{data}
\name{Activity}
\alias{Activity}
\title{Activity}
\format{\code{\link{R6Class}} object, super class \code{\link{PersistentObject}}.}
\usage{
Activity
}
\description{
Activity
}
\section{Fields}{

\describe{
\item{\code{isDeleted}}{of type bool inherited from super class \code{\link{PersistentObject}}.}

\item{\code{rev}}{of type String inherited from super class \code{\link{PersistentObject}}.}

\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{type}}{of type String.}

\item{\code{objectKind}}{of type String.}

\item{\code{teamId}}{of type String.}

\item{\code{projectId}}{of type String.}

\item{\code{userId}}{of type String.}

\item{\code{projectName}}{of type String.}

\item{\code{isPublic}}{of type bool.}

\item{\code{date}}{object of class \code{\link{Date}}.}

\item{\code{properties}}{list of class \code{\link{Pair}}.}
}}

\keyword{datasets}
