% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_count.R
\docType{data}
\name{ActivityCount}
\alias{ActivityCount}
\title{ActivityCount}
\format{\code{\link{R6Class}} object.}
\usage{
ActivityCount
}
\description{
ActivityCount
}
\section{Fields}{

\describe{
\item{\code{objectId}}{of type String.}

\item{\code{count}}{of type int.}
}}

\keyword{datasets}
