% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_service.R
\docType{data}
\name{ActivityService}
\alias{ActivityService}
\title{ActivityService}
\format{\code{\link{R6Class}} object.}
\usage{
ActivityService
}
\description{
ActivityService
}
\section{Methods}{

\describe{
   \item{\code{getPublicActivityCount(kind,limit)}}{method}
}
}

\keyword{datasets}
