% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\docType{data}
\name{Address}
\alias{Address}
\title{Address}
\format{\code{\link{R6Class}} object.}
\usage{
Address
}
\description{
Address
}
\section{Fields}{

\describe{
\item{\code{country}}{of type String.}

\item{\code{state}}{of type String.}

\item{\code{city}}{of type String.}

\item{\code{zipCode}}{of type String.}

\item{\code{address1}}{of type String.}

\item{\code{address2}}{of type String.}

\item{\code{phone}}{of type String.}
}}

\keyword{datasets}
