% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_call_profile.R
\docType{data}
\name{ApiCallProfile}
\alias{ApiCallProfile}
\title{ApiCallProfile}
\format{\code{\link{R6Class}} object, super class \code{\link{Profile}}.}
\usage{
ApiCallProfile
}
\description{
ApiCallProfile
}
\section{Fields}{

\describe{
\item{\code{name}}{of type String inherited from super class \code{\link{Profile}}.}

\item{\code{nCalls}}{of type int.}
}}

\keyword{datasets}
