% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute.R
\docType{data}
\name{Attribute}
\alias{Attribute}
\title{Attribute}
\format{\code{\link{R6Class}} object, super class \code{\link{Factor}}.}
\usage{
Attribute
}
\description{
Attribute
}
\section{Fields}{

\describe{
\item{\code{name}}{of type String inherited from super class \code{\link{Factor}}.}

\item{\code{type}}{of type String inherited from super class \code{\link{Factor}}.}

\item{\code{relationId}}{of type String.}
}}

\keyword{datasets}
