% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\docType{data}
\name{Axis}
\alias{Axis}
\title{Axis}
\format{\code{\link{R6Class}} object.}
\usage{
Axis
}
\description{
Axis
}
\section{Fields}{

\describe{
\item{\code{axisExtent}}{object of class \code{\link{Point}}.}

\item{\code{axisSettings}}{object of class \code{\link{AxisSettings}}.}

\item{\code{graphicalFactor}}{object of class \code{\link{GraphicalFactor}}.}
}}

\keyword{datasets}
