% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis_settings.R
\docType{data}
\name{AxisSettings}
\alias{AxisSettings}
\title{AxisSettings}
\format{\code{\link{R6Class}} object.}
\usage{
AxisSettings
}
\description{
AxisSettings
}
\section{Fields}{

\describe{
\item{\code{properties}}{list of class \code{\link{Property}}.}

\item{\code{propertyValues}}{list of class \code{\link{PropertyValue}}.}
}}

\keyword{datasets}
