% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billing_info.R
\docType{data}
\name{BillingInfo}
\alias{BillingInfo}
\title{BillingInfo}
\format{\code{\link{R6Class}} object.}
\usage{
BillingInfo
}
\description{
BillingInfo
}
\section{Fields}{

\describe{
\item{\code{firstName}}{of type String.}

\item{\code{lastName}}{of type String.}

\item{\code{companyName}}{of type String.}

\item{\code{taxId}}{object of class \code{\link{TaxId}}.}

\item{\code{address}}{object of class \code{\link{Address}}.}
}}

\keyword{datasets}
