% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boolean_property.R
\docType{data}
\name{BooleanProperty}
\alias{BooleanProperty}
\title{BooleanProperty}
\format{\code{\link{R6Class}} object, super class \code{\link{Property}}.}
\usage{
BooleanProperty
}
\description{
BooleanProperty
}
\section{Fields}{

\describe{
\item{\code{name}}{of type String inherited from super class \code{\link{Property}}.}
}}

\keyword{datasets}
