% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canceled_state.R
\docType{data}
\name{CanceledState}
\alias{CanceledState}
\title{CanceledState}
\format{\code{\link{R6Class}} object, super class \code{\link{State}}.}
\usage{
CanceledState
}
\description{
CanceledState
}
\keyword{datasets}
