% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/category_palette.R
\docType{data}
\name{CategoryPalette}
\alias{CategoryPalette}
\title{CategoryPalette}
\format{\code{\link{R6Class}} object, super class \code{\link{Palette}}.}
\usage{
CategoryPalette
}
\description{
CategoryPalette
}
\section{Fields}{

\describe{
\item{\code{backcolor}}{of type int inherited from super class \code{\link{Palette}}.}

\item{\code{colorList}}{object of class \code{\link{ColorList}}.}

\item{\code{stringColorElements}}{list of class \code{\link{StringColorElement}}.}
}}

\keyword{datasets}
