% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.R
\docType{data}
\name{Chart}
\alias{Chart}
\title{Chart}
\format{\code{\link{R6Class}} object, sub classes \code{\link{ChartLine}}, \code{\link{ChartPoint}}, \code{\link{ChartHeatmap}}, \code{\link{ChartBar}}, \code{\link{ChartSize}}.}
\usage{
Chart
}
\description{
Chart
}
\section{Fields}{

\describe{
\item{\code{name}}{of type String.}

\item{\code{properties}}{object of class \code{\link{Properties}}.}
}}

\keyword{datasets}
