% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_bar.R
\docType{data}
\name{ChartBar}
\alias{ChartBar}
\title{ChartBar}
\format{\code{\link{R6Class}} object, super class \code{\link{Chart}}.}
\usage{
ChartBar
}
\description{
ChartBar
}
\section{Fields}{

\describe{
\item{\code{name}}{of type String inherited from super class \code{\link{Chart}}.}

\item{\code{properties}}{object of class \code{\link{Properties}} inherited from super class \code{\link{Chart}}.}
}}

\keyword{datasets}
