% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_heatmap.R
\docType{data}
\name{ChartHeatmap}
\alias{ChartHeatmap}
\title{ChartHeatmap}
\format{\code{\link{R6Class}} object, super class \code{\link{Chart}}.}
\usage{
ChartHeatmap
}
\description{
ChartHeatmap
}
\keyword{datasets}
