% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_line.R
\docType{data}
\name{ChartLine}
\alias{ChartLine}
\title{ChartLine}
\format{\code{\link{R6Class}} object, super class \code{\link{ChartSize}}.}
\usage{
ChartLine
}
\description{
ChartLine
}
\section{Fields}{

\describe{
\item{\code{pointSize}}{of type int inherited from super class \code{\link{ChartSize}}.}

\item{\code{name}}{of type String inherited from super class \code{\link{Chart}}.}

\item{\code{properties}}{object of class \code{\link{Properties}} inherited from super class \code{\link{Chart}}.}
}}

\keyword{datasets}
