% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_point.R
\docType{data}
\name{ChartPoint}
\alias{ChartPoint}
\title{ChartPoint}
\format{\code{\link{R6Class}} object, super class \code{\link{ChartSize}}.}
\usage{
ChartPoint
}
\description{
ChartPoint
}
\section{Fields}{

\describe{
\item{\code{pointSize}}{of type int inherited from super class \code{\link{ChartSize}}.}

\item{\code{name}}{of type String inherited from super class \code{\link{Chart}}.}

\item{\code{properties}}{object of class \code{\link{Properties}} inherited from super class \code{\link{Chart}}.}
}}

\keyword{datasets}
