% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_size.R
\docType{data}
\name{ChartSize}
\alias{ChartSize}
\title{ChartSize}
\format{\code{\link{R6Class}} object, super class \code{\link{Chart}}, sub classes \code{\link{ChartLine}}, \code{\link{ChartPoint}}.}
\usage{
ChartSize
}
\description{
ChartSize
}
\section{Fields}{

\describe{
\item{\code{name}}{of type String inherited from super class \code{\link{Chart}}.}

\item{\code{pointSize}}{of type int.}

\item{\code{properties}}{object of class \code{\link{Properties}} inherited from super class \code{\link{Chart}}.}
}}

\keyword{datasets}
