% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_element.R
\docType{data}
\name{ColorElement}
\alias{ColorElement}
\title{ColorElement}
\format{\code{\link{R6Class}} object, sub classes \code{\link{DoubleColorElement}}, \code{\link{StringColorElement}}.}
\usage{
ColorElement
}
\description{
ColorElement
}
\section{Fields}{

\describe{
\item{\code{color}}{of type int.}
}}

\keyword{datasets}
