% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_hash_palette.R
\docType{data}
\name{ColorHashPalette}
\alias{ColorHashPalette}
\title{ColorHashPalette}
\format{\code{\link{R6Class}} object, super class \code{\link{CategoryPalette}}.}
\usage{
ColorHashPalette
}
\description{
ColorHashPalette
}
\section{Fields}{

\describe{
\item{\code{backcolor}}{of type int inherited from super class \code{\link{Palette}}.}

\item{\code{colorList}}{object of class \code{\link{ColorList}} inherited from super class \code{\link{CategoryPalette}}.}

\item{\code{stringColorElements}}{list of class \code{\link{StringColorElement}} inherited from super class \code{\link{CategoryPalette}}.}
}}

\keyword{datasets}
