% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{Colors}
\alias{Colors}
\title{Colors}
\format{\code{\link{R6Class}} object.}
\usage{
Colors
}
\description{
Colors
}
\section{Fields}{

\describe{
\item{\code{factors}}{list of class \code{\link{Factor}}.}

\item{\code{palette}}{object of class \code{\link{Palette}}.}
}}

\keyword{datasets}
