% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column.R
\docType{data}
\name{Column}
\alias{Column}
\title{Column}
\format{\code{\link{R6Class}} object, super class \code{\link{ColumnSchema}}.}
\usage{
Column
}
\description{
Column
}
\section{Fields}{

\describe{
\item{\code{name}}{of type String inherited from super class \code{\link{ColumnSchema}}.}

\item{\code{type}}{of type String inherited from super class \code{\link{ColumnSchema}}.}

\item{\code{nRows}}{of type int inherited from super class \code{\link{ColumnSchema}}.}

\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{values}}{of type dynamic.}

\item{\code{metaData}}{object of class \code{\link{ColumnSchemaMetaData}} inherited from super class \code{\link{ColumnSchema}}.}
}}

\keyword{datasets}
