% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_pair.R
\docType{data}
\name{ColumnPair}
\alias{ColumnPair}
\title{ColumnPair}
\format{\code{\link{R6Class}} object.}
\usage{
ColumnPair
}
\description{
ColumnPair
}
\section{Fields}{

\describe{
\item{\code{lColumns}}{list of type String.}

\item{\code{rColumns}}{list of type String.}
}}

\keyword{datasets}
