% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_schema.R
\docType{data}
\name{ColumnSchema}
\alias{ColumnSchema}
\title{ColumnSchema}
\format{\code{\link{R6Class}} object, super class \code{\link{IdObject}}, sub classes \code{\link{Column}}.}
\usage{
ColumnSchema
}
\description{
ColumnSchema
}
\section{Fields}{

\describe{
\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{name}}{of type String.}

\item{\code{type}}{of type String.}

\item{\code{nRows}}{of type int.}

\item{\code{metaData}}{object of class \code{\link{ColumnSchemaMetaData}}.}
}}

\keyword{datasets}
