% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_relation.R
\docType{data}
\name{CompositeRelation}
\alias{CompositeRelation}
\title{CompositeRelation}
\format{\code{\link{R6Class}} object, super class \code{\link{Relation}}.}
\usage{
CompositeRelation
}
\description{
CompositeRelation
}
\section{Fields}{

\describe{
\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{joinOperators}}{list of class \code{\link{JoinOperator}}.}

\item{\code{mainRelation}}{object of class \code{\link{Relation}}.}
}}

\keyword{datasets}
