% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computation_task.R
\docType{data}
\name{ComputationTask}
\alias{ComputationTask}
\title{ComputationTask}
\format{\code{\link{R6Class}} object, super class \code{\link{CubeQueryTask}}.}
\usage{
ComputationTask
}
\description{
ComputationTask
}
\section{Fields}{

\describe{
\item{\code{removeOnGC}}{of type bool inherited from super class \code{\link{CubeQueryTask}}.}

\item{\code{schemaIds}}{list of type String inherited from super class \code{\link{CubeQueryTask}}.}

\item{\code{projectId}}{of type String inherited from super class \code{\link{ProjectTask}}.}

\item{\code{duration}}{of type double inherited from super class \code{\link{Task}}.}

\item{\code{owner}}{of type String inherited from super class \code{\link{Task}}.}

\item{\code{taskHash}}{of type String inherited from super class \code{\link{Task}}.}

\item{\code{isDeleted}}{of type bool inherited from super class \code{\link{PersistentObject}}.}

\item{\code{rev}}{of type String inherited from super class \code{\link{PersistentObject}}.}

\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{parentTaskId}}{of type String.}

\item{\code{fileResultId}}{of type String.}

\item{\code{query}}{object of class \code{\link{CubeQuery}} inherited from super class \code{\link{CubeQueryTask}}.}

\item{\code{environment}}{list of class \code{\link{Pair}} inherited from super class \code{\link{Task}}.}

\item{\code{state}}{object of class \code{\link{State}} inherited from super class \code{\link{Task}}.}

\item{\code{createdDate}}{object of class \code{\link{Date}} inherited from super class \code{\link{Task}}.}

\item{\code{lastModifiedDate}}{object of class \code{\link{Date}} inherited from super class \code{\link{Task}}.}

\item{\code{runDate}}{object of class \code{\link{Date}} inherited from super class \code{\link{Task}}.}

\item{\code{completedDate}}{object of class \code{\link{Date}} inherited from super class \code{\link{Task}}.}

\item{\code{aclContext}}{object of class \code{\link{AclContext}} inherited from super class \code{\link{Task}}.}

\item{\code{computedRelation}}{object of class \code{\link{Relation}}.}
}}

\keyword{datasets}
