% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpu_time_profile.R
\docType{data}
\name{CpuTimeProfile}
\alias{CpuTimeProfile}
\title{CpuTimeProfile}
\format{\code{\link{R6Class}} object, super class \code{\link{Profile}}.}
\usage{
CpuTimeProfile
}
\description{
CpuTimeProfile
}
\section{Fields}{

\describe{
\item{\code{name}}{of type String inherited from super class \code{\link{Profile}}.}

\item{\code{cpuTime}}{of type double.}
}}

\keyword{datasets}
