% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_library_service.R
\name{CranLibraryService}
\alias{CranLibraryService}
\title{CranLibraryService}
\format{\code{\link{R6Class}} object.}
\description{
CranLibraryService

CranLibraryService
}
\section{Methods}{

\describe{
   \item{\code{packagesGz(repoName)}}{method}
   \item{\code{packagesRds(repoName)}}{method}
   \item{\code{packages(repoName)}}{method}
   \item{\code{package(repoName,package)}}{method}
}
}

\section{Super class}{
\code{\link[tercen:HttpClientService]{tercen::HttpClientService}} -> \code{CranLibraryService}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CranLibraryService$new()}}
\item \href{#method-findByOwner}{\code{CranLibraryService$findByOwner()}}
\item \href{#method-findByFileId}{\code{CranLibraryService$findByFileId()}}
\item \href{#method-findByOwnerNameVersion}{\code{CranLibraryService$findByOwnerNameVersion()}}
\item \href{#method-packagesGz}{\code{CranLibraryService$packagesGz()}}
\item \href{#method-packagesRds}{\code{CranLibraryService$packagesRds()}}
\item \href{#method-packages}{\code{CranLibraryService$packages()}}
\item \href{#method-package}{\code{CranLibraryService$package()}}
\item \href{#method-clone}{\code{CranLibraryService$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="create">}\href{../../tercen/html/HttpClientService.html#method-create}{\code{tercen::HttpClientService$create()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="delete">}\href{../../tercen/html/HttpClientService.html#method-delete}{\code{tercen::HttpClientService$delete()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="findKeys">}\href{../../tercen/html/HttpClientService.html#method-findKeys}{\code{tercen::HttpClientService$findKeys()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="findStartKeys">}\href{../../tercen/html/HttpClientService.html#method-findStartKeys}{\code{tercen::HttpClientService$findStartKeys()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="fromTson">}\href{../../tercen/html/HttpClientService.html#method-fromTson}{\code{tercen::HttpClientService$fromTson()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="get">}\href{../../tercen/html/HttpClientService.html#method-get}{\code{tercen::HttpClientService$get()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="getServiceUri">}\href{../../tercen/html/HttpClientService.html#method-getServiceUri}{\code{tercen::HttpClientService$getServiceUri()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="list">}\href{../../tercen/html/HttpClientService.html#method-list}{\code{tercen::HttpClientService$list()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="onResponseError">}\href{../../tercen/html/HttpClientService.html#method-onResponseError}{\code{tercen::HttpClientService$onResponseError()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="toTson">}\href{../../tercen/html/HttpClientService.html#method-toTson}{\code{tercen::HttpClientService$toTson()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="update">}\href{../../tercen/html/HttpClientService.html#method-update}{\code{tercen::HttpClientService$update()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CranLibraryService$new(baseRestUri, client)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-findByOwner"></a>}}
\subsection{Method \code{findByOwner()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CranLibraryService$findByOwner(keys = NULL, useFactory = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-findByFileId"></a>}}
\subsection{Method \code{findByFileId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CranLibraryService$findByFileId(keys = NULL, useFactory = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-findByOwnerNameVersion"></a>}}
\subsection{Method \code{findByOwnerNameVersion()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CranLibraryService$findByOwnerNameVersion(
  startKey = NULL,
  endKey = NULL,
  limit = 20,
  skip = 0,
  descending = TRUE,
  useFactory = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-packagesGz"></a>}}
\subsection{Method \code{packagesGz()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CranLibraryService$packagesGz(repoName)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-packagesRds"></a>}}
\subsection{Method \code{packagesRds()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CranLibraryService$packagesRds(repoName)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-packages"></a>}}
\subsection{Method \code{packages()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CranLibraryService$packages(repoName)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-package"></a>}}
\subsection{Method \code{package()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CranLibraryService$package(repoName, package)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CranLibraryService$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
