% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_git_operator_task.R
\docType{data}
\name{CreateGitOperatorTask}
\alias{CreateGitOperatorTask}
\title{CreateGitOperatorTask}
\format{\code{\link{R6Class}} object, super class \code{\link{Task}}.}
\usage{
CreateGitOperatorTask
}
\description{
CreateGitOperatorTask
}
\section{Fields}{

\describe{
\item{\code{duration}}{of type double inherited from super class \code{\link{Task}}.}

\item{\code{owner}}{of type String inherited from super class \code{\link{Task}}.}

\item{\code{taskHash}}{of type String inherited from super class \code{\link{Task}}.}

\item{\code{isDeleted}}{of type bool inherited from super class \code{\link{PersistentObject}}.}

\item{\code{rev}}{of type String inherited from super class \code{\link{PersistentObject}}.}

\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{version}}{of type String.}

\item{\code{operatorId}}{of type String.}

\item{\code{environment}}{list of class \code{\link{Pair}} inherited from super class \code{\link{Task}}.}

\item{\code{state}}{object of class \code{\link{State}} inherited from super class \code{\link{Task}}.}

\item{\code{createdDate}}{object of class \code{\link{Date}} inherited from super class \code{\link{Task}}.}

\item{\code{lastModifiedDate}}{object of class \code{\link{Date}} inherited from super class \code{\link{Task}}.}

\item{\code{runDate}}{object of class \code{\link{Date}} inherited from super class \code{\link{Task}}.}

\item{\code{completedDate}}{object of class \code{\link{Date}} inherited from super class \code{\link{Task}}.}

\item{\code{aclContext}}{object of class \code{\link{AclContext}} inherited from super class \code{\link{Task}}.}

\item{\code{url}}{object of class \code{\link{Url}}.}
}}

\keyword{datasets}
