% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstab.R
\docType{data}
\name{Crosstab}
\alias{Crosstab}
\title{Crosstab}
\format{\code{\link{R6Class}} object, super class \code{\link{StepModel}}.}
\usage{
Crosstab
}
\description{
Crosstab
}
\section{Fields}{

\describe{
\item{\code{taskId}}{of type String.}

\item{\code{axis}}{object of class \code{\link{XYAxisList}}.}

\item{\code{columnTable}}{object of class \code{\link{CrosstabTable}}.}

\item{\code{filters}}{object of class \code{\link{Filters}}.}

\item{\code{operatorSettings}}{object of class \code{\link{OperatorSettings}}.}

\item{\code{rowTable}}{object of class \code{\link{CrosstabTable}}.}
}}

\keyword{datasets}
