% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_tab_step.R
\docType{data}
\name{CrossTabStep}
\alias{CrossTabStep}
\title{CrossTabStep}
\format{\code{\link{R6Class}} object, super class \code{\link{NamespaceStep}}, sub classes \code{\link{DataStep}}, \code{\link{ViewDataStep}}.}
\usage{
CrossTabStep
}
\description{
CrossTabStep
}
\section{Fields}{

\describe{
\item{\code{namespace}}{of type String inherited from super class \code{\link{NamespaceStep}}.}

\item{\code{groupId}}{of type String inherited from super class \code{\link{Step}}.}

\item{\code{name}}{of type String inherited from super class \code{\link{Step}}.}

\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{model}}{object of class \code{\link{StepModel}} inherited from super class \code{\link{ModelStep}}.}

\item{\code{inputs}}{list of class \code{\link{InputPort}} inherited from super class \code{\link{Step}}.}

\item{\code{outputs}}{list of class \code{\link{OutputPort}} inherited from super class \code{\link{Step}}.}

\item{\code{rectangle}}{object of class \code{\link{Rectangle}} inherited from super class \code{\link{Step}}.}

\item{\code{state}}{object of class \code{\link{StepState}} inherited from super class \code{\link{Step}}.}

\item{\code{model}}{object of class \code{\link{StepModel}} inherited from super class \code{\link{Step}}.}

\item{\code{model}}{object of class \code{\link{Crosstab}}.}
}}

\keyword{datasets}
