% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstab_table.R
\docType{data}
\name{CrosstabTable}
\alias{CrosstabTable}
\title{CrosstabTable}
\format{\code{\link{R6Class}} object.}
\usage{
CrosstabTable
}
\description{
CrosstabTable
}
\section{Fields}{

\describe{
\item{\code{cellSize}}{of type double.}

\item{\code{offset}}{of type int.}

\item{\code{graphicalFactors}}{list of class \code{\link{GraphicalFactor}}.}

\item{\code{rectangleSelections}}{list of class \code{\link{Rectangle}}.}
}}

\keyword{datasets}
