% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_s_v_file_metadata.R
\docType{data}
\name{CSVFileMetadata}
\alias{CSVFileMetadata}
\title{CSVFileMetadata}
\format{\code{\link{R6Class}} object, super class \code{\link{FileMetadata}}.}
\usage{
CSVFileMetadata
}
\description{
CSVFileMetadata
}
\section{Fields}{

\describe{
\item{\code{contentType}}{of type String inherited from super class \code{\link{FileMetadata}}.}

\item{\code{cacheControl}}{of type String inherited from super class \code{\link{FileMetadata}}.}

\item{\code{contentEncoding}}{of type String inherited from super class \code{\link{FileMetadata}}.}

\item{\code{contentLanguage}}{of type String inherited from super class \code{\link{FileMetadata}}.}

\item{\code{md5Hash}}{of type String inherited from super class \code{\link{FileMetadata}}.}

\item{\code{separator}}{of type String.}

\item{\code{quote}}{of type String.}

\item{\code{headers}}{of type bool.}
}}

\keyword{datasets}
