% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_s_v_parser_param.R
\docType{data}
\name{CSVParserParam}
\alias{CSVParserParam}
\title{CSVParserParam}
\format{\code{\link{R6Class}} object.}
\usage{
CSVParserParam
}
\description{
CSVParserParam
}
\section{Fields}{

\describe{
\item{\code{separator}}{of type String.}

\item{\code{encoding}}{of type String.}

\item{\code{quote}}{of type String.}

\item{\code{hasHeaders}}{of type bool.}

\item{\code{allowMalformed}}{of type bool.}

\item{\code{comment}}{of type String.}
}}

\keyword{datasets}
