% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_s_v_task.R
\docType{data}
\name{CSVTask}
\alias{CSVTask}
\title{CSVTask}
\format{\code{\link{R6Class}} object, super class \code{\link{Task}}.}
\usage{
CSVTask
}
\description{
CSVTask
}
\section{Fields}{

\describe{
\item{\code{projectId}}{of type String inherited from super class \code{\link{Task}}.}

\item{\code{taskHash}}{of type String inherited from super class \code{\link{Task}}.}

\item{\code{isDeleted}}{of type bool inherited from super class \code{\link{PersistentObject}}.}

\item{\code{rev}}{of type String inherited from super class \code{\link{PersistentObject}}.}

\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{fileDocumentId}}{of type String.}

\item{\code{schemaId}}{of type String.}

\item{\code{state}}{object of class \code{\link{State}} inherited from super class \code{\link{Task}}.}

\item{\code{runDate}}{object of class \code{\link{Date}} inherited from super class \code{\link{Task}}.}

\item{\code{completedDate}}{object of class \code{\link{Date}} inherited from super class \code{\link{Task}}.}

\item{\code{aclContext}}{object of class \code{\link{AclContext}} inherited from super class \code{\link{Task}}.}
}}

\keyword{datasets}
