% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube_axis_query.R
\docType{data}
\name{CubeAxisQuery}
\alias{CubeAxisQuery}
\title{CubeAxisQuery}
\format{\code{\link{R6Class}} object.}
\usage{
CubeAxisQuery
}
\description{
CubeAxisQuery
}
\section{Fields}{

\describe{
\item{\code{pointSize}}{of type int.}

\item{\code{chartType}}{of type String.}

\item{\code{yAxis}}{object of class \code{\link{Factor}}.}

\item{\code{yAxisSettings}}{object of class \code{\link{AxisSettings}}.}

\item{\code{xAxis}}{object of class \code{\link{Factor}}.}

\item{\code{xAxisSettings}}{object of class \code{\link{AxisSettings}}.}

\item{\code{errors}}{list of class \code{\link{Factor}}.}

\item{\code{labels}}{list of class \code{\link{Factor}}.}

\item{\code{colors}}{list of class \code{\link{Factor}}.}
}}

\keyword{datasets}
