% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube_query.R
\docType{data}
\name{CubeQuery}
\alias{CubeQuery}
\title{CubeQuery}
\format{\code{\link{R6Class}} object.}
\usage{
CubeQuery
}
\description{
CubeQuery
}
\section{Fields}{

\describe{
\item{\code{qtHash}}{of type String.}

\item{\code{columnHash}}{of type String.}

\item{\code{rowHash}}{of type String.}

\item{\code{relation}}{object of class \code{\link{Relation}}.}

\item{\code{colColumns}}{list of class \code{\link{Factor}}.}

\item{\code{rowColumns}}{list of class \code{\link{Factor}}.}

\item{\code{axisQueries}}{list of class \code{\link{CubeAxisQuery}}.}

\item{\code{filters}}{object of class \code{\link{Filters}}.}

\item{\code{operatorSettings}}{object of class \code{\link{OperatorSettings}}.}
}}

\keyword{datasets}
