% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube_query_table_schema.R
\docType{data}
\name{CubeQueryTableSchema}
\alias{CubeQueryTableSchema}
\title{CubeQueryTableSchema}
\format{\code{\link{R6Class}} object, super class \code{\link{TableSchema}}.}
\usage{
CubeQueryTableSchema
}
\description{
CubeQueryTableSchema
}
\section{Fields}{

\describe{
\item{\code{nRows}}{of type int inherited from super class \code{\link{TableSchema}}.}

\item{\code{dataDirectory}}{of type String inherited from super class \code{\link{TableSchema}}.}

\item{\code{projectId}}{of type String inherited from super class \code{\link{ProjectDocument}}.}

\item{\code{description}}{of type String inherited from super class \code{\link{Document}}.}

\item{\code{name}}{of type String inherited from super class \code{\link{Document}}.}

\item{\code{tags}}{list of type String inherited from super class \code{\link{Document}}.}

\item{\code{isDeleted}}{of type bool inherited from super class \code{\link{PersistentObject}}.}

\item{\code{rev}}{of type String inherited from super class \code{\link{PersistentObject}}.}

\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{queryHash}}{of type String.}

\item{\code{columns}}{list of class \code{\link{ColumnSchema}} inherited from super class \code{\link{TableSchema}}.}

\item{\code{acl}}{object of class \code{\link{Acl}} inherited from super class \code{\link{Document}}.}

\item{\code{createdDate}}{object of class \code{\link{Date}} inherited from super class \code{\link{Document}}.}

\item{\code{lastModifiedDate}}{object of class \code{\link{Date}} inherited from super class \code{\link{Document}}.}

\item{\code{urls}}{list of class \code{\link{Url}} inherited from super class \code{\link{Document}}.}
}}

\keyword{datasets}
