% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description.R
\docType{data}
\name{Description}
\alias{Description}
\title{Description}
\format{\code{\link{R6Class}} object.}
\usage{
Description
}
\description{
Description
}
\section{Fields}{

\describe{
\item{\code{Package}}{of type String.}

\item{\code{Version}}{of type String.}

\item{\code{Depends}}{of type String.}

\item{\code{Imports}}{of type String.}

\item{\code{LinkingTo}}{of type String.}

\item{\code{Suggests}}{of type String.}

\item{\code{License}}{of type String.}

\item{\code{MD5sum}}{of type String.}

\item{\code{NeedsCompilation}}{of type String.}
}}

\keyword{datasets}
