% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distinct_relation.R
\docType{data}
\name{DistinctRelation}
\alias{DistinctRelation}
\title{DistinctRelation}
\format{\code{\link{R6Class}} object, super class \code{\link{Relation}}.}
\usage{
DistinctRelation
}
\description{
DistinctRelation
}
\section{Fields}{

\describe{
\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{group}}{list of type String.}

\item{\code{relation}}{object of class \code{\link{Relation}}.}
}}

\keyword{datasets}
