% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\docType{data}
\name{Document}
\alias{Document}
\title{Document}
\format{\code{\link{R6Class}} object, super class \code{\link{PersistentObject}}, sub classes \code{\link{Team}}, \code{\link{ShinyOperator}}, \code{\link{ROperator}}, \code{\link{WebAppOperator}}, \code{\link{GitOperator}}, \code{\link{CubeQueryTableSchema}}, \code{\link{TableSchema}}, \code{\link{ComputedTableSchema}}, \code{\link{Schema}}, \code{\link{FileDocument}}, \code{\link{Workflow}}, \code{\link{User}}, \code{\link{Operator}}, \code{\link{ProjectDocument}}, \code{\link{Project}}.}
\usage{
Document
}
\description{
Document
}
\section{Fields}{

\describe{
\item{\code{isDeleted}}{of type bool inherited from super class \code{\link{PersistentObject}}.}

\item{\code{rev}}{of type String inherited from super class \code{\link{PersistentObject}}.}

\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{description}}{of type String.}

\item{\code{name}}{of type String.}

\item{\code{createdBy}}{of type String.}

\item{\code{tags}}{list of type String.}

\item{\code{version}}{of type String.}

\item{\code{authors}}{list of type String.}

\item{\code{acl}}{object of class \code{\link{Acl}}.}

\item{\code{createdDate}}{object of class \code{\link{Date}}.}

\item{\code{lastModifiedDate}}{object of class \code{\link{Date}}.}

\item{\code{urls}}{list of class \code{\link{Url}}.}

\item{\code{meta}}{list of class \code{\link{Pair}}.}

\item{\code{url}}{object of class \code{\link{Url}}.}
}}

\keyword{datasets}
