% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_service.R
\docType{data}
\name{DocumentService}
\alias{DocumentService}
\title{DocumentService}
\format{\code{\link{R6Class}} object.}
\usage{
DocumentService
}
\description{
DocumentService
}
\section{Methods}{

\describe{
   \item{\code{getTercenOperatorLibrary(offset,limit)}}{method}
   \item{\code{getTercenWorkflowLibrary(offset,limit)}}{method}
   \item{\code{getTercenAppLibrary(offset,limit)}}{method}
}
}

\keyword{datasets}
