% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_step_model.R
\docType{data}
\name{DocumentStepModel}
\alias{DocumentStepModel}
\title{DocumentStepModel}
\format{\code{\link{R6Class}} object, super class \code{\link{StepModel}}.}
\usage{
DocumentStepModel
}
\description{
DocumentStepModel
}
\section{Fields}{

\describe{
\item{\code{documentIds}}{list of type String.}
}}

\keyword{datasets}
