% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_color_element.R
\docType{data}
\name{DoubleColorElement}
\alias{DoubleColorElement}
\title{DoubleColorElement}
\format{\code{\link{R6Class}} object, super class \code{\link{ColorElement}}.}
\usage{
DoubleColorElement
}
\description{
DoubleColorElement
}
\section{Fields}{

\describe{
\item{\code{color}}{of type int inherited from super class \code{\link{ColorElement}}.}

\item{\code{doubleValue}}{of type double.}
}}

\keyword{datasets}
