% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_property.R
\docType{data}
\name{DoubleProperty}
\alias{DoubleProperty}
\title{DoubleProperty}
\format{\code{\link{R6Class}} object, super class \code{\link{Property}}.}
\usage{
DoubleProperty
}
\description{
DoubleProperty
}
\section{Fields}{

\describe{
\item{\code{name}}{of type String inherited from super class \code{\link{Property}}.}
}}

\keyword{datasets}
