% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumerated_property.R
\docType{data}
\name{EnumeratedProperty}
\alias{EnumeratedProperty}
\title{EnumeratedProperty}
\format{\code{\link{R6Class}} object, super class \code{\link{Property}}.}
\usage{
EnumeratedProperty
}
\description{
EnumeratedProperty
}
\section{Fields}{

\describe{
\item{\code{name}}{of type String inherited from super class \code{\link{Property}}.}
}}

\keyword{datasets}
