% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\docType{data}
\name{Event}
\alias{Event}
\title{Event}
\format{\code{\link{R6Class}} object, super class \code{\link{PersistentObject}}, sub classes \code{\link{TaskStateEvent}}, \code{\link{TaskLogEvent}}, \code{\link{TaskProgressEvent}}, \code{\link{TaskDataEvent}}, \code{\link{TaskEvent}}.}
\usage{
Event
}
\description{
Event
}
\section{Fields}{

\describe{
\item{\code{isDeleted}}{of type bool inherited from super class \code{\link{PersistentObject}}.}

\item{\code{rev}}{of type String inherited from super class \code{\link{PersistentObject}}.}

\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{date}}{object of class \code{\link{Date}}.}
}}

\keyword{datasets}
