% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_service.R
\docType{data}
\name{EventService}
\alias{EventService}
\title{EventService}
\format{\code{\link{R6Class}} object.}
\usage{
EventService
}
\description{
EventService
}
\section{Methods}{

\describe{
   \item{\code{sendChannel(channel,evt)}}{method}
   \item{\code{channel(name)}}{method}
   \item{\code{listenTaskChannel(taskId,start)}}{method}
   \item{\code{onTaskState(taskId)}}{method}
   \item{\code{taskListenerCount(taskId)}}{method}
}
}

\keyword{datasets}
