% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_model.R
\docType{data}
\name{ExportModel}
\alias{ExportModel}
\title{ExportModel}
\format{\code{\link{R6Class}} object, super class \code{\link{StepModel}}.}
\usage{
ExportModel
}
\description{
ExportModel
}
\keyword{datasets}
