% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_operator.R
\docType{data}
\name{ExternalOperator}
\alias{ExternalOperator}
\title{ExternalOperator}
\format{\code{\link{R6Class}} object, super class \code{\link{Operator}}.}
\usage{
ExternalOperator
}
\description{
ExternalOperator
}
\section{Fields}{

\describe{
\item{\code{projectId}}{of type String inherited from super class \code{\link{ProjectDocument}}.}

\item{\code{description}}{of type String inherited from super class \code{\link{Document}}.}

\item{\code{name}}{of type String inherited from super class \code{\link{Document}}.}

\item{\code{tags}}{list of type String inherited from super class \code{\link{Document}}.}

\item{\code{isDeleted}}{of type bool inherited from super class \code{\link{PersistentObject}}.}

\item{\code{rev}}{of type String inherited from super class \code{\link{PersistentObject}}.}

\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{properties}}{list of class \code{\link{Property}} inherited from super class \code{\link{Operator}}.}

\item{\code{acl}}{object of class \code{\link{Acl}} inherited from super class \code{\link{Document}}.}

\item{\code{createdDate}}{object of class \code{\link{Date}} inherited from super class \code{\link{Document}}.}

\item{\code{lastModifiedDate}}{object of class \code{\link{Date}} inherited from super class \code{\link{Document}}.}

\item{\code{urls}}{list of class \code{\link{Url}} inherited from super class \code{\link{Document}}.}

\item{\code{url}}{object of class \code{\link{Url}}.}
}}

\keyword{datasets}
