% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.R
\docType{data}
\name{Factor}
\alias{Factor}
\title{Factor}
\format{\code{\link{R6Class}} object, sub classes \code{\link{Attribute}}, \code{\link{MappingFactor}}.}
\usage{
Factor
}
\description{
Factor
}
\section{Fields}{

\describe{
\item{\code{name}}{of type String.}

\item{\code{type}}{of type String.}
}}

\keyword{datasets}
