% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/failed_state.R
\docType{data}
\name{FailedState}
\alias{FailedState}
\title{FailedState}
\format{\code{\link{R6Class}} object, super class \code{\link{State}}.}
\usage{
FailedState
}
\description{
FailedState
}
\section{Fields}{

\describe{
\item{\code{error}}{of type String.}

\item{\code{reason}}{of type String.}
}}

\keyword{datasets}
