% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialisation.R
\name{file_to_tercen}
\alias{file_to_tercen}
\title{Convert File to Tercen Table}
\usage{
file_to_tercen(file_path, chunk_size_bits = 1e+06, filename = NULL)
}
\description{
Serialise a file to a Tercen operator output table.
}
