% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_metadata.R
\docType{data}
\name{FileMetadata}
\alias{FileMetadata}
\title{FileMetadata}
\format{\code{\link{R6Class}} object, sub classes \code{\link{CSVFileMetadata}}.}
\usage{
FileMetadata
}
\description{
FileMetadata
}
\section{Fields}{

\describe{
\item{\code{contentType}}{of type String.}

\item{\code{cacheControl}}{of type String.}

\item{\code{contentEncoding}}{of type String.}

\item{\code{contentLanguage}}{of type String.}

\item{\code{md5Hash}}{of type String.}
}}

\keyword{datasets}
