% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_service.R
\docType{data}
\name{FileService}
\alias{FileService}
\title{FileService}
\format{\code{\link{R6Class}} object.}
\usage{
FileService
}
\description{
FileService
}
\section{Methods}{

\describe{
   \item{\code{upload(file,bytes)}}{method}
   \item{\code{download(fileDocumentId)}}{method}
}
}

\keyword{datasets}
