% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\docType{data}
\name{Filter}
\alias{Filter}
\title{Filter}
\format{\code{\link{R6Class}} object, super class \code{\link{FilterTopExpr}}, sub classes \code{\link{NamedFilter}}.}
\usage{
Filter
}
\description{
Filter
}
\section{Fields}{

\describe{
\item{\code{logical}}{of type String.}

\item{\code{not}}{of type bool.}

\item{\code{filterExprs}}{list of class \code{\link{FilterTopExpr}}.}
}}

\keyword{datasets}
