% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_expr.R
\docType{data}
\name{FilterExpr}
\alias{FilterExpr}
\title{FilterExpr}
\format{\code{\link{R6Class}} object, super class \code{\link{FilterTopExpr}}.}
\usage{
FilterExpr
}
\description{
FilterExpr
}
\section{Fields}{

\describe{
\item{\code{filterOp}}{of type String.}

\item{\code{stringValue}}{of type String.}

\item{\code{factor}}{object of class \code{\link{Factor}}.}
}}

\keyword{datasets}
