% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\docType{data}
\name{Filters}
\alias{Filters}
\title{Filters}
\format{\code{\link{R6Class}} object.}
\usage{
Filters
}
\description{
Filters
}
\section{Fields}{

\describe{
\item{\code{removeNaN}}{of type bool.}

\item{\code{namedFilters}}{list of class \code{\link{NamedFilter}}.}
}}

\keyword{datasets}
