% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_top_expr.R
\docType{data}
\name{FilterTopExpr}
\alias{FilterTopExpr}
\title{FilterTopExpr}
\format{\code{\link{R6Class}} object, sub classes \code{\link{NamedFilter}}, \code{\link{Filter}}, \code{\link{FilterExpr}}.}
\usage{
FilterTopExpr
}
\description{
FilterTopExpr
}
\keyword{datasets}
