% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder_service.R
\docType{data}
\name{FolderService}
\alias{FolderService}
\title{FolderService}
\format{\code{\link{R6Class}} object.}
\usage{
FolderService
}
\description{
FolderService
}
\section{Methods}{

\describe{
   \item{\code{getOrCreate(projectId,path)}}{method}
}
}

\keyword{datasets}
