% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_property.R
\docType{data}
\name{FormulaProperty}
\alias{FormulaProperty}
\title{FormulaProperty}
\format{\code{\link{R6Class}} object, super class \code{\link{StringProperty}}.}
\usage{
FormulaProperty
}
\description{
FormulaProperty
}
\section{Fields}{

\describe{
\item{\code{defaultValue}}{of type String inherited from super class \code{\link{StringProperty}}.}

\item{\code{name}}{of type String inherited from super class \code{\link{Property}}.}

\item{\code{description}}{of type String inherited from super class \code{\link{Property}}.}
}}

\keyword{datasets}
