% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garbage_collector_service.R
\docType{data}
\name{GarbageCollectorService}
\alias{GarbageCollectorService}
\title{GarbageCollectorService}
\format{\code{\link{R6Class}} object.}
\usage{
GarbageCollectorService
}
\description{
GarbageCollectorService
}
\section{Methods}{

\describe{
}
}

\keyword{datasets}
