% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garbage_object.R
\docType{data}
\name{GarbageObject}
\alias{GarbageObject}
\title{GarbageObject}
\format{\code{\link{R6Class}} object, super class \code{\link{PersistentObject}}, sub classes \code{\link{GarbageTasks}}.}
\usage{
GarbageObject
}
\description{
GarbageObject
}
\section{Fields}{

\describe{
\item{\code{isDeleted}}{of type bool inherited from super class \code{\link{PersistentObject}}.}

\item{\code{rev}}{of type String inherited from super class \code{\link{PersistentObject}}.}

\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}
}}

\keyword{datasets}
