% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garbage_tasks.R
\docType{data}
\name{GarbageTasks}
\alias{GarbageTasks}
\title{GarbageTasks}
\format{\code{\link{R6Class}} object, super class \code{\link{GarbageObject}}.}
\usage{
GarbageTasks
}
\description{
GarbageTasks
}
\section{Fields}{

\describe{
\item{\code{isDeleted}}{of type bool inherited from super class \code{\link{PersistentObject}}.}

\item{\code{rev}}{of type String inherited from super class \code{\link{PersistentObject}}.}

\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{workflowId}}{of type String.}

\item{\code{deletedTaskIds}}{list of type String.}

\item{\code{addedTaskIds}}{list of type String.}
}}

\keyword{datasets}
