% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_relation.R
\docType{data}
\name{GatherRelation}
\alias{GatherRelation}
\title{GatherRelation}
\format{\code{\link{R6Class}} object, super class \code{\link{Relation}}.}
\usage{
GatherRelation
}
\description{
GatherRelation
}
\section{Fields}{

\describe{
\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{names}}{list of type String.}

\item{\code{valueName}}{of type String.}

\item{\code{variableName}}{of type String.}

\item{\code{valueType}}{of type String.}

\item{\code{relation}}{object of class \code{\link{Relation}}.}
}}

\keyword{datasets}
