% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_event.R
\docType{data}
\name{GenericEvent}
\alias{GenericEvent}
\title{GenericEvent}
\format{\code{\link{R6Class}} object, super class \code{\link{Event}}.}
\usage{
GenericEvent
}
\description{
GenericEvent
}
\section{Fields}{

\describe{
\item{\code{isDeleted}}{of type bool inherited from super class \code{\link{PersistentObject}}.}

\item{\code{rev}}{of type String inherited from super class \code{\link{PersistentObject}}.}

\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{type}}{of type String.}

\item{\code{content}}{of type String.}

\item{\code{date}}{object of class \code{\link{Date}} inherited from super class \code{\link{Event}}.}
}}

\keyword{datasets}
