% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical_factor.R
\docType{data}
\name{GraphicalFactor}
\alias{GraphicalFactor}
\title{GraphicalFactor}
\format{\code{\link{R6Class}} object.}
\usage{
GraphicalFactor
}
\description{
GraphicalFactor
}
\section{Fields}{

\describe{
\item{\code{factor}}{object of class \code{\link{Factor}}.}

\item{\code{rectangle}}{object of class \code{\link{Rectangle}}.}
}}

\keyword{datasets}
