% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_step.R
\docType{data}
\name{GroupStep}
\alias{GroupStep}
\title{GroupStep}
\format{\code{\link{R6Class}} object, super class \code{\link{RelationStep}}.}
\usage{
GroupStep
}
\description{
GroupStep
}
\section{Fields}{

\describe{
\item{\code{groupId}}{of type String inherited from super class \code{\link{Step}}.}

\item{\code{name}}{of type String inherited from super class \code{\link{Step}}.}

\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{appId}}{of type String.}

\item{\code{appName}}{of type String.}

\item{\code{version}}{of type String.}

\item{\code{inputs}}{list of class \code{\link{InputPort}} inherited from super class \code{\link{Step}}.}

\item{\code{outputs}}{list of class \code{\link{OutputPort}} inherited from super class \code{\link{Step}}.}

\item{\code{rectangle}}{object of class \code{\link{Rectangle}} inherited from super class \code{\link{Step}}.}

\item{\code{state}}{object of class \code{\link{StepState}} inherited from super class \code{\link{Step}}.}

\item{\code{offset}}{object of class \code{\link{Point}}.}
}}

\keyword{datasets}
