% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_memory_relation.R
\docType{data}
\name{InMemoryRelation}
\alias{InMemoryRelation}
\title{InMemoryRelation}
\format{\code{\link{R6Class}} object, super class \code{\link{Relation}}.}
\usage{
InMemoryRelation
}
\description{
InMemoryRelation
}
\section{Fields}{

\describe{
\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}
}}

\keyword{datasets}
