% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_port.R
\docType{data}
\name{InputPort}
\alias{InputPort}
\title{InputPort}
\format{\code{\link{R6Class}} object, super class \code{\link{Port}}.}
\usage{
InputPort
}
\description{
InputPort
}
\section{Fields}{

\describe{
\item{\code{linkType}}{of type String inherited from super class \code{\link{Port}}.}

\item{\code{name}}{of type String inherited from super class \code{\link{Port}}.}

\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}
}}

\keyword{datasets}
