% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_step.R
\docType{data}
\name{InStep}
\alias{InStep}
\title{InStep}
\format{\code{\link{R6Class}} object, super class \code{\link{RelationStep}}.}
\usage{
InStep
}
\description{
InStep
}
\section{Fields}{

\describe{
\item{\code{groupId}}{of type String inherited from super class \code{\link{Step}}.}

\item{\code{name}}{of type String inherited from super class \code{\link{Step}}.}

\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{inputs}}{list of class \code{\link{InputPort}} inherited from super class \code{\link{Step}}.}

\item{\code{outputs}}{list of class \code{\link{OutputPort}} inherited from super class \code{\link{Step}}.}

\item{\code{rectangle}}{object of class \code{\link{Rectangle}} inherited from super class \code{\link{Step}}.}

\item{\code{state}}{object of class \code{\link{StepState}} inherited from super class \code{\link{Step}}.}

\item{\code{groupPortPosition}}{object of class \code{\link{Point}}.}
}}

\keyword{datasets}
