% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integer_property.R
\docType{data}
\name{IntegerProperty}
\alias{IntegerProperty}
\title{IntegerProperty}
\format{\code{\link{R6Class}} object, super class \code{\link{Property}}.}
\usage{
IntegerProperty
}
\description{
IntegerProperty
}
\section{Fields}{

\describe{
\item{\code{name}}{of type String inherited from super class \code{\link{Property}}.}
}}

\keyword{datasets}
