% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issue_message_service.R
\docType{data}
\name{IssueMessageService}
\alias{IssueMessageService}
\title{IssueMessageService}
\format{\code{\link{R6Class}} object.}
\usage{
IssueMessageService
}
\description{
IssueMessageService
}
\section{Methods}{

\describe{
}
}

\keyword{datasets}
