% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issue_service.R
\docType{data}
\name{IssueService}
\alias{IssueService}
\title{IssueService}
\format{\code{\link{R6Class}} object.}
\usage{
IssueService
}
\description{
IssueService
}
\section{Methods}{

\describe{
}
}

\keyword{datasets}
