% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jet_palette.R
\docType{data}
\name{JetPalette}
\alias{JetPalette}
\title{JetPalette}
\format{\code{\link{R6Class}} object, super class \code{\link{RampPalette}}.}
\usage{
JetPalette
}
\description{
JetPalette
}
\section{Fields}{

\describe{
\item{\code{isUserDefined}}{of type bool inherited from super class \code{\link{RampPalette}}.}

\item{\code{backcolor}}{of type int inherited from super class \code{\link{Palette}}.}

\item{\code{doubleColorElements}}{list of class \code{\link{DoubleColorElement}} inherited from super class \code{\link{RampPalette}}.}
}}

\keyword{datasets}
