% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_melt_operator.R
\docType{data}
\name{JoinMeltOperator}
\alias{JoinMeltOperator}
\title{JoinMeltOperator}
\format{\code{\link{R6Class}} object, super class \code{\link{JoinOperator}}.}
\usage{
JoinMeltOperator
}
\description{
JoinMeltOperator
}
\section{Fields}{

\describe{
\item{\code{rightRelationId}}{of type String.}

\item{\code{names}}{list of type String.}

\item{\code{valueName}}{of type String.}

\item{\code{variableName}}{of type String.}

\item{\code{valueType}}{of type String.}
}}

\keyword{datasets}
