% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_operator.R
\docType{data}
\name{JoinOperator}
\alias{JoinOperator}
\title{JoinOperator}
\format{\code{\link{R6Class}} object.}
\usage{
JoinOperator
}
\description{
JoinOperator
}
\section{Fields}{

\describe{
\item{\code{leftPair}}{object of class \code{\link{ColumnPair}}.}

\item{\code{rightRelation}}{object of class \code{\link{Relation}}.}
}}

\keyword{datasets}
