% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_operator1.R
\docType{data}
\name{JoinOperator1}
\alias{JoinOperator1}
\title{JoinOperator1}
\format{\code{\link{R6Class}} object, super class \code{\link{JoinOperator}}.}
\usage{
JoinOperator1
}
\description{
JoinOperator1
}
\section{Fields}{

\describe{
\item{\code{leftPair}}{object of class \code{\link{ColumnPair}}.}

\item{\code{rightRelation}}{object of class \code{\link{Relation}}.}
}}

\keyword{datasets}
