% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_step_model.R
\docType{data}
\name{JoinStepModel}
\alias{JoinStepModel}
\title{JoinStepModel}
\format{\code{\link{R6Class}} object, super class \code{\link{StepModel}}.}
\usage{
JoinStepModel
}
\description{
JoinStepModel
}
\section{Fields}{

\describe{
\item{\code{rightPrefix}}{of type String.}

\item{\code{leftFactors}}{list of class \code{\link{Factor}}.}

\item{\code{rightFactors}}{list of class \code{\link{Factor}}.}
}}

\keyword{datasets}
