% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\docType{data}
\name{Labels}
\alias{Labels}
\title{Labels}
\format{\code{\link{R6Class}} object.}
\usage{
Labels
}
\description{
Labels
}
\section{Fields}{

\describe{
\item{\code{factors}}{list of class \code{\link{Factor}}.}
}}

\keyword{datasets}
