% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\docType{data}
\name{Library}
\alias{Library}
\title{Library}
\format{\code{\link{R6Class}} object, super class \code{\link{PersistentObject}}, sub classes \code{\link{RenvInstalledLibrary}}.}
\usage{
Library
}
\description{
Library
}
\section{Fields}{

\describe{
\item{\code{isDeleted}}{of type bool inherited from super class \code{\link{PersistentObject}}.}

\item{\code{rev}}{of type String inherited from super class \code{\link{PersistentObject}}.}

\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{description}}{object of class \code{\link{Description}}.}
}}

\keyword{datasets}
