% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\docType{data}
\name{Link}
\alias{Link}
\title{Link}
\format{\code{\link{R6Class}} object, super class \code{\link{IdObject}}.}
\usage{
Link
}
\description{
Link
}
\section{Fields}{

\describe{
\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{inputId}}{of type String.}

\item{\code{outputId}}{of type String.}
}}

\keyword{datasets}
