% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lock.R
\docType{data}
\name{Lock}
\alias{Lock}
\title{Lock}
\format{\code{\link{R6Class}} object, super class \code{\link{PersistentObject}}.}
\usage{
Lock
}
\description{
Lock
}
\section{Fields}{

\describe{
\item{\code{isDeleted}}{of type bool inherited from super class \code{\link{PersistentObject}}.}

\item{\code{rev}}{of type String inherited from super class \code{\link{PersistentObject}}.}

\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{name}}{of type String.}
}}

\keyword{datasets}
