% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lock_service.R
\docType{data}
\name{LockService}
\alias{LockService}
\title{LockService}
\format{\code{\link{R6Class}} object.}
\usage{
LockService
}
\description{
LockService
}
\section{Methods}{

\describe{
   \item{\code{lock(name,wait)}}{method}
   \item{\code{releaseLock(lock)}}{method}
}
}

\keyword{datasets}
