% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_factor.R
\docType{data}
\name{MappingFactor}
\alias{MappingFactor}
\title{MappingFactor}
\format{\code{\link{R6Class}} object, super class \code{\link{Factor}}.}
\usage{
MappingFactor
}
\description{
MappingFactor
}
\section{Fields}{

\describe{
\item{\code{name}}{of type String inherited from super class \code{\link{Factor}}.}

\item{\code{type}}{of type String inherited from super class \code{\link{Factor}}.}

\item{\code{isSingle}}{of type bool.}

\item{\code{description}}{of type String.}

\item{\code{factorName}}{of type String.}

\item{\code{factors}}{list of class \code{\link{Factor}}.}
}}

\keyword{datasets}
