% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_filter.R
\docType{data}
\name{MappingFilter}
\alias{MappingFilter}
\title{MappingFilter}
\format{\code{\link{R6Class}} object.}
\usage{
MappingFilter
}
\description{
MappingFilter
}
\section{Fields}{

\describe{
\item{\code{name}}{of type String.}

\item{\code{description}}{of type String.}

\item{\code{namedFilter}}{object of class \code{\link{NamedFilter}}.}
}}

\keyword{datasets}
