% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt_step.R
\docType{data}
\name{MeltStep}
\alias{MeltStep}
\title{MeltStep}
\format{\code{\link{R6Class}} object, super class \code{\link{NamespaceStep}}.}
\usage{
MeltStep
}
\description{
MeltStep
}
\section{Fields}{

\describe{
\item{\code{groupId}}{of type String inherited from super class \code{\link{Step}}.}

\item{\code{name}}{of type String inherited from super class \code{\link{Step}}.}

\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{inputs}}{list of class \code{\link{InputPort}} inherited from super class \code{\link{Step}}.}

\item{\code{outputs}}{list of class \code{\link{OutputPort}} inherited from super class \code{\link{Step}}.}

\item{\code{rectangle}}{object of class \code{\link{Rectangle}} inherited from super class \code{\link{Step}}.}

\item{\code{state}}{object of class \code{\link{StepState}} inherited from super class \code{\link{Step}}.}

\item{\code{model}}{object of class \code{\link{MeltStepModel}}.}

\item{\code{meltedAttributes}}{list of class \code{\link{Attribute}}.}
}}

\keyword{datasets}
