% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named_filter.R
\docType{data}
\name{NamedFilter}
\alias{NamedFilter}
\title{NamedFilter}
\format{\code{\link{R6Class}} object, super class \code{\link{Filter}}.}
\usage{
NamedFilter
}
\description{
NamedFilter
}
\section{Fields}{

\describe{
\item{\code{logical}}{of type String inherited from super class \code{\link{Filter}}.}

\item{\code{not}}{of type bool inherited from super class \code{\link{Filter}}.}

\item{\code{name}}{of type String.}

\item{\code{filterExprs}}{list of class \code{\link{FilterTopExpr}} inherited from super class \code{\link{Filter}}.}
}}

\keyword{datasets}
