% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator_ref.R
\docType{data}
\name{OperatorRef}
\alias{OperatorRef}
\title{OperatorRef}
\format{\code{\link{R6Class}} object.}
\usage{
OperatorRef
}
\description{
OperatorRef
}
\section{Fields}{

\describe{
\item{\code{name}}{of type String.}

\item{\code{operatorId}}{of type String.}

\item{\code{operatorKind}}{of type String.}

\item{\code{version}}{of type String.}

\item{\code{propertyValues}}{list of class \code{\link{PropertyValue}}.}
}}

\keyword{datasets}
