% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator_result.R
\docType{data}
\name{OperatorResult}
\alias{OperatorResult}
\title{OperatorResult}
\format{\code{\link{R6Class}} object.}
\usage{
OperatorResult
}
\description{
OperatorResult
}
\section{Fields}{

\describe{
\item{\code{tables}}{list of class \code{\link{Table}}.}

\item{\code{joinOperators}}{list of class \code{\link{JoinOperator}}.}
}}

\keyword{datasets}
