% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator_service.R
\docType{data}
\name{OperatorService}
\alias{OperatorService}
\title{OperatorService}
\format{\code{\link{R6Class}} object.}
\usage{
OperatorService
}
\description{
OperatorService
}
\section{Methods}{

\describe{
}
}

\keyword{datasets}
