% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator_system.R
\docType{data}
\name{OperatorSystem}
\alias{OperatorSystem}
\title{OperatorSystem}
\format{\code{\link{R6Class}} object.}
\usage{
OperatorSystem
}
\description{
OperatorSystem
}
\section{Fields}{

\describe{
\item{\code{name}}{of type String.}

\item{\code{memory}}{of type int.}

\item{\code{kernelMemory}}{of type int.}

\item{\code{blkioWeight}}{of type int.}

\item{\code{pidsLimit}}{of type int.}

\item{\code{ulimits_nofile}}{of type int.}

\item{\code{timeout}}{of type int.}

\item{\code{storageSize}}{of type String.}

\item{\code{cpusetCpus}}{of type String.}
}}

\keyword{datasets}
