% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator_table_result.R
\docType{data}
\name{OperatorTableResult}
\alias{OperatorTableResult}
\title{OperatorTableResult}
\format{\code{\link{R6Class}} object, super class \code{\link{OperatorResult}}.}
\usage{
OperatorTableResult
}
\description{
OperatorTableResult
}
\section{Fields}{

\describe{
\item{\code{joinOperators}}{list of class \code{\link{JoinOperator}} inherited from super class \code{\link{OperatorResult}}.}

\item{\code{tables}}{list of class \code{\link{Table}}.}
}}

\keyword{datasets}
