% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator_unit_test.R
\docType{data}
\name{OperatorUnitTest}
\alias{OperatorUnitTest}
\title{OperatorUnitTest}
\format{\code{\link{R6Class}} object.}
\usage{
OperatorUnitTest
}
\description{
OperatorUnitTest
}
\section{Fields}{

\describe{
\item{\code{name}}{of type String.}

\item{\code{namespace}}{of type String.}

\item{\code{inputDataUri}}{of type String.}

\item{\code{outputDataUri}}{list of type String.}

\item{\code{columns}}{list of type String.}

\item{\code{rows}}{list of type String.}

\item{\code{colors}}{list of type String.}

\item{\code{labels}}{list of type String.}

\item{\code{yAxis}}{of type String.}

\item{\code{xAxis}}{of type String.}
}}

\keyword{datasets}
