% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_port.R
\docType{data}
\name{OutputPort}
\alias{OutputPort}
\title{OutputPort}
\format{\code{\link{R6Class}} object, super class \code{\link{Port}}.}
\usage{
OutputPort
}
\description{
OutputPort
}
\section{Fields}{

\describe{
\item{\code{linkType}}{of type String inherited from super class \code{\link{Port}}.}

\item{\code{name}}{of type String inherited from super class \code{\link{Port}}.}

\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}
}}

\keyword{datasets}
