% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair.R
\docType{data}
\name{Pair}
\alias{Pair}
\title{Pair}
\format{\code{\link{R6Class}} object.}
\usage{
Pair
}
\description{
Pair
}
\section{Fields}{

\describe{
\item{\code{key}}{of type String.}

\item{\code{value}}{of type String.}
}}

\keyword{datasets}
