% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\docType{data}
\name{Palette}
\alias{Palette}
\title{Palette}
\format{\code{\link{R6Class}} object, sub classes \code{\link{JetPalette}}, \code{\link{RampPalette}}, \code{\link{CategoryPalette}}.}
\usage{
Palette
}
\description{
Palette
}
\section{Fields}{

\describe{
\item{\code{backcolor}}{of type int.}

\item{\code{properties}}{list of class \code{\link{PropertyValue}}.}
}}

\keyword{datasets}
