% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistent_service.R
\docType{data}
\name{PersistentService}
\alias{PersistentService}
\title{PersistentService}
\format{\code{\link{R6Class}} object.}
\usage{
PersistentService
}
\description{
PersistentService
}
\section{Methods}{

\describe{
   \item{\code{summary(teamOrProjectId)}}{method}
}
}

\keyword{datasets}
