% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\docType{data}
\name{Plan}
\alias{Plan}
\title{Plan}
\format{\code{\link{R6Class}} object.}
\usage{
Plan
}
\description{
Plan
}
\section{Fields}{

\describe{
\item{\code{name}}{of type String.}

\item{\code{displayName}}{of type String.}

\item{\code{paymentProviderPlanId}}{of type String.}

\item{\code{descriptions}}{list of type String.}

\item{\code{price}}{of type double.}
}}

\keyword{datasets}
