% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point.R
\docType{data}
\name{Point}
\alias{Point}
\title{Point}
\format{\code{\link{R6Class}} object.}
\usage{
Point
}
\description{
Point
}
\section{Fields}{

\describe{
\item{\code{x}}{of type double.}

\item{\code{y}}{of type double.}
}}

\keyword{datasets}
