% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port.R
\docType{data}
\name{Port}
\alias{Port}
\title{Port}
\format{\code{\link{R6Class}} object, super class \code{\link{IdObject}}, sub classes \code{\link{InputPort}}, \code{\link{OutputPort}}.}
\usage{
Port
}
\description{
Port
}
\section{Fields}{

\describe{
\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{linkType}}{of type String.}

\item{\code{name}}{of type String.}
}}

\keyword{datasets}
