% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/privilege.R
\docType{data}
\name{Privilege}
\alias{Privilege}
\title{Privilege}
\format{\code{\link{R6Class}} object.}
\usage{
Privilege
}
\description{
Privilege
}
\section{Fields}{

\describe{
\item{\code{type}}{of type String.}
}}

\keyword{datasets}
