% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\docType{data}
\name{Profile}
\alias{Profile}
\title{Profile}
\format{\code{\link{R6Class}} object, sub classes \code{\link{StorageProfile}}, \code{\link{RunProfile}}, \code{\link{CpuTimeProfile}}, \code{\link{TableProfile}}, \code{\link{ApiCallProfile}}.}
\usage{
Profile
}
\description{
Profile
}
\section{Fields}{

\describe{
\item{\code{name}}{of type String.}
}}

\keyword{datasets}
