% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\docType{data}
\name{Profiles}
\alias{Profiles}
\title{Profiles}
\format{\code{\link{R6Class}} object.}
\usage{
Profiles
}
\description{
Profiles
}
\section{Fields}{

\describe{
\item{\code{apiProfile}}{object of class \code{\link{ApiCallProfile}}.}

\item{\code{tableProfile}}{object of class \code{\link{TableProfile}}.}

\item{\code{cpuTimeProfile}}{object of class \code{\link{CpuTimeProfile}}.}

\item{\code{storageProfile}}{object of class \code{\link{StorageProfile}}.}

\item{\code{runProfile}}{object of class \code{\link{RunProfile}}.}
}}

\keyword{datasets}
