% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_document_service.R
\docType{data}
\name{ProjectDocumentService}
\alias{ProjectDocumentService}
\title{ProjectDocumentService}
\format{\code{\link{R6Class}} object.}
\usage{
ProjectDocumentService
}
\description{
ProjectDocumentService
}
\section{Methods}{

\describe{
   \item{\code{getParentFolders(documentId)}}{method}
}
}

\keyword{datasets}
