% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_task.R
\docType{data}
\name{ProjectTask}
\alias{ProjectTask}
\title{ProjectTask}
\format{\code{\link{R6Class}} object, super class \code{\link{Task}}, sub classes \code{\link{ComputationTask}}, \code{\link{RunComputationTask}}, \code{\link{SaveComputationResultTask}}, \code{\link{ExportWorkflowTask}}, \code{\link{RunWebAppTask}}, \code{\link{CSVTask}}, \code{\link{CubeQueryTask}}, \code{\link{ExportTableTask}}, \code{\link{ImportWorkflowTask}}, \code{\link{ImportGitWorkflowTask}}.}
\usage{
ProjectTask
}
\description{
ProjectTask
}
\section{Fields}{

\describe{
\item{\code{duration}}{of type double inherited from super class \code{\link{Task}}.}

\item{\code{owner}}{of type String inherited from super class \code{\link{Task}}.}

\item{\code{taskHash}}{of type String inherited from super class \code{\link{Task}}.}

\item{\code{runProfile}}{of type String inherited from super class \code{\link{Task}}.}

\item{\code{isDeleted}}{of type bool inherited from super class \code{\link{PersistentObject}}.}

\item{\code{rev}}{of type String inherited from super class \code{\link{PersistentObject}}.}

\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{projectId}}{of type String.}

\item{\code{state}}{object of class \code{\link{State}} inherited from super class \code{\link{Task}}.}

\item{\code{createdDate}}{object of class \code{\link{Date}} inherited from super class \code{\link{Task}}.}

\item{\code{lastModifiedDate}}{object of class \code{\link{Date}} inherited from super class \code{\link{Task}}.}

\item{\code{runDate}}{object of class \code{\link{Date}} inherited from super class \code{\link{Task}}.}

\item{\code{completedDate}}{object of class \code{\link{Date}} inherited from super class \code{\link{Task}}.}

\item{\code{aclContext}}{object of class \code{\link{AclContext}} inherited from super class \code{\link{Task}}.}
}}

\keyword{datasets}
