% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\docType{data}
\name{Properties}
\alias{Properties}
\title{Properties}
\format{\code{\link{R6Class}} object.}
\usage{
Properties
}
\description{
Properties
}
\section{Fields}{

\describe{
\item{\code{properties}}{list of class \code{\link{Property}}.}

\item{\code{propertyValues}}{list of class \code{\link{PropertyValue}}.}
}}

\keyword{datasets}
