% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/property.R
\docType{data}
\name{Property}
\alias{Property}
\title{Property}
\format{\code{\link{R6Class}} object, sub classes \code{\link{EnumeratedProperty}}, \code{\link{DoubleProperty}}, \code{\link{StringProperty}}, \code{\link{BooleanProperty}}.}
\usage{
Property
}
\description{
Property
}
\section{Fields}{

\describe{
\item{\code{name}}{of type String.}
}}

\keyword{datasets}
