% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/property_value.R
\docType{data}
\name{PropertyValue}
\alias{PropertyValue}
\title{PropertyValue}
\format{\code{\link{R6Class}} object.}
\usage{
PropertyValue
}
\description{
PropertyValue
}
\section{Fields}{

\describe{
\item{\code{name}}{of type String.}

\item{\code{value}}{of type Object.}
}}

\keyword{datasets}
