% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\docType{data}
\name{Proxy}
\alias{Proxy}
\title{Proxy}
\format{\code{\link{R6Class}} object.}
\usage{
Proxy
}
\description{
Proxy
}
\section{Fields}{

\describe{
\item{\code{name}}{of type String.}

\item{\code{targetUrls}}{list of type String.}
}}

\keyword{datasets}
