% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramp_palette.R
\docType{data}
\name{RampPalette}
\alias{RampPalette}
\title{RampPalette}
\format{\code{\link{R6Class}} object, super class \code{\link{Palette}}, sub classes \code{\link{JetPalette}}.}
\usage{
RampPalette
}
\description{
RampPalette
}
\section{Fields}{

\describe{
\item{\code{backcolor}}{of type int inherited from super class \code{\link{Palette}}.}

\item{\code{isUserDefined}}{of type bool.}

\item{\code{properties}}{list of class \code{\link{PropertyValue}} inherited from super class \code{\link{Palette}}.}

\item{\code{doubleColorElements}}{list of class \code{\link{DoubleColorElement}}.}
}}

\keyword{datasets}
