% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectangle.R
\docType{data}
\name{Rectangle}
\alias{Rectangle}
\title{Rectangle}
\format{\code{\link{R6Class}} object.}
\usage{
Rectangle
}
\description{
Rectangle
}
\section{Fields}{

\describe{
\item{\code{extent}}{object of class \code{\link{Point}}.}

\item{\code{topLeft}}{object of class \code{\link{Point}}.}
}}

\keyword{datasets}
