% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_relation.R
\docType{data}
\name{ReferenceRelation}
\alias{ReferenceRelation}
\title{ReferenceRelation}
\format{\code{\link{R6Class}} object, super class \code{\link{SimpleRelation}}.}
\usage{
ReferenceRelation
}
\description{
ReferenceRelation
}
\section{Fields}{

\describe{
\item{\code{index}}{of type int inherited from super class \code{\link{SimpleRelation}}.}

\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{relation}}{object of class \code{\link{Relation}}.}
}}

\keyword{datasets}
