% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_relation.R
\docType{data}
\name{RenameRelation}
\alias{RenameRelation}
\title{RenameRelation}
\format{\code{\link{R6Class}} object, super class \code{\link{Relation}}.}
\usage{
RenameRelation
}
\description{
RenameRelation
}
\section{Fields}{

\describe{
\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{inNames}}{list of type String.}

\item{\code{outNames}}{list of type String.}

\item{\code{relation}}{object of class \code{\link{Relation}}.}
}}

\keyword{datasets}
