% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_summary.R
\docType{data}
\name{ResourceSummary}
\alias{ResourceSummary}
\title{ResourceSummary}
\format{\code{\link{R6Class}} object.}
\usage{
ResourceSummary
}
\description{
ResourceSummary
}
\section{Fields}{

\describe{
\item{\code{storage}}{of type int.}

\item{\code{usedStorage}}{of type int.}

\item{\code{cpuTime}}{of type double.}

\item{\code{usedCpuTime}}{of type double.}
}}

\keyword{datasets}
