% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_proxy.R
\docType{data}
\name{RProxy}
\alias{RProxy}
\title{RProxy}
\format{\code{\link{R6Class}} object.}
\usage{
RProxy
}
\description{
RProxy
}
\section{Fields}{

\describe{
\item{\code{name}}{of type String.}

\item{\code{targetUrl}}{of type String.}
}}

\keyword{datasets}
