% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_source_library.R
\docType{data}
\name{RSourceLibrary}
\alias{RSourceLibrary}
\title{RSourceLibrary}
\format{\code{\link{R6Class}} object, super class \code{\link{RLibrary}}.}
\usage{
RSourceLibrary
}
\description{
RSourceLibrary
}
\section{Fields}{

\describe{
\item{\code{description}}{of type String inherited from super class \code{\link{Document}}.}

\item{\code{name}}{of type String inherited from super class \code{\link{Document}}.}

\item{\code{createdBy}}{of type String inherited from super class \code{\link{Document}}.}

\item{\code{tags}}{list of type String inherited from super class \code{\link{Document}}.}

\item{\code{version}}{of type String inherited from super class \code{\link{Document}}.}

\item{\code{authors}}{list of type String inherited from super class \code{\link{Document}}.}

\item{\code{isPublic}}{of type bool inherited from super class \code{\link{Document}}.}

\item{\code{isDeleted}}{of type bool inherited from super class \code{\link{PersistentObject}}.}

\item{\code{rev}}{of type String inherited from super class \code{\link{PersistentObject}}.}

\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{fileId}}{of type String.}

\item{\code{rDescription}}{object of class \code{\link{RDescription}} inherited from super class \code{\link{RLibrary}}.}

\item{\code{acl}}{object of class \code{\link{Acl}} inherited from super class \code{\link{Document}}.}

\item{\code{createdDate}}{object of class \code{\link{Date}} inherited from super class \code{\link{Document}}.}

\item{\code{lastModifiedDate}}{object of class \code{\link{Date}} inherited from super class \code{\link{Document}}.}

\item{\code{urls}}{list of class \code{\link{Url}} inherited from super class \code{\link{Document}}.}

\item{\code{meta}}{list of class \code{\link{Pair}} inherited from super class \code{\link{Document}}.}

\item{\code{url}}{object of class \code{\link{Url}} inherited from super class \code{\link{Document}}.}
}}

\keyword{datasets}
