% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/running_dependent_state.R
\docType{data}
\name{RunningDependentState}
\alias{RunningDependentState}
\title{RunningDependentState}
\format{\code{\link{R6Class}} object, super class \code{\link{State}}.}
\usage{
RunningDependentState
}
\description{
RunningDependentState
}
\keyword{datasets}
