% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/running_environment.R
\docType{data}
\name{RunningEnvironment}
\alias{RunningEnvironment}
\title{RunningEnvironment}
\format{\code{\link{R6Class}} object, sub classes \code{\link{WorkerEndpointRunningEnvironment}}, \code{\link{SimpleRunningEnvironment}}.}
\usage{
RunningEnvironment
}
\description{
RunningEnvironment
}
\keyword{datasets}
