% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/running_state.R
\docType{data}
\name{RunningState}
\alias{RunningState}
\title{RunningState}
\format{\code{\link{R6Class}} object, super class \code{\link{State}}.}
\usage{
RunningState
}
\description{
RunningState
}
\keyword{datasets}
