% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_profile.R
\docType{data}
\name{RunProfile}
\alias{RunProfile}
\title{RunProfile}
\format{\code{\link{R6Class}} object.}
\usage{
RunProfile
}
\description{
RunProfile
}
\section{Fields}{

\describe{
\item{\code{name}}{of type String.}

\item{\code{memory}}{of type int.}

\item{\code{kernelMemory}}{of type int.}

\item{\code{blkioWeight}}{of type int.}

\item{\code{pidsLimit}}{of type int.}

\item{\code{ulimits_nofile}}{of type int.}

\item{\code{timeout}}{of type int.}

\item{\code{storageSize}}{of type String.}

\item{\code{cpusetCpus}}{of type String.}
}}

\keyword{datasets}
