% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_running_environment.R
\docType{data}
\name{SimpleRunningEnvironment}
\alias{SimpleRunningEnvironment}
\title{SimpleRunningEnvironment}
\format{\code{\link{R6Class}} object, super class \code{\link{RunningEnvironment}}.}
\usage{
SimpleRunningEnvironment
}
\description{
SimpleRunningEnvironment
}
\section{Fields}{

\describe{
\item{\code{cpus}}{of type int.}

\item{\code{memory}}{of type double.}
}}

\keyword{datasets}
