% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_task_resource.R
\docType{data}
\name{SimpleTaskResource}
\alias{SimpleTaskResource}
\title{SimpleTaskResource}
\format{\code{\link{R6Class}} object, super class \code{\link{RunningEnvironment}}.}
\usage{
SimpleTaskResource
}
\description{
SimpleTaskResource
}
\section{Fields}{

\describe{
\item{\code{cpus}}{of type int.}

\item{\code{memory}}{of type double.}
}}

\keyword{datasets}
