% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_process.R
\docType{data}
\name{StartProcess}
\alias{StartProcess}
\title{StartProcess}
\format{\code{\link{R6Class}} object, super class \code{\link{IdObject}}.}
\usage{
StartProcess
}
\description{
StartProcess
}
\section{Fields}{

\describe{
\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{executable}}{of type String.}

\item{\code{arguments}}{list of type String.}

\item{\code{timeout}}{of type int.}

\item{\code{pid}}{of type int.}

\item{\code{script}}{of type String.}

\item{\code{ulimits}}{object of class \code{\link{Ulimits}}.}
}}

\keyword{datasets}
