% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\docType{data}
\name{State}
\alias{State}
\title{State}
\format{\code{\link{R6Class}} object, sub classes \code{\link{RunningState}}, \code{\link{FailedState}}, \code{\link{CanceledState}}, \code{\link{DoneState}}, \code{\link{InitState}}.}
\usage{
State
}
\description{
State
}
\keyword{datasets}
