% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\docType{data}
\name{Step}
\alias{Step}
\title{Step}
\format{\code{\link{R6Class}} object, super class \code{\link{IdObject}}, sub classes \code{\link{DataStep}}, \code{\link{MeltStep}}, \code{\link{JoinStep}}, \code{\link{WizardStep}}, \code{\link{CrossTabStep}}, \code{\link{InStep}}, \code{\link{GroupStep}}, \code{\link{OutStep}}, \code{\link{TableStep}}, \code{\link{NamespaceStep}}, \code{\link{RelationStep}}, \code{\link{ModelStep}}, \code{\link{ViewStep}}.}
\usage{
Step
}
\description{
Step
}
\section{Fields}{

\describe{
\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{groupId}}{of type String.}

\item{\code{name}}{of type String.}

\item{\code{inputs}}{list of class \code{\link{InputPort}}.}

\item{\code{outputs}}{list of class \code{\link{OutputPort}}.}

\item{\code{rectangle}}{object of class \code{\link{Rectangle}}.}

\item{\code{state}}{object of class \code{\link{StepState}}.}
}}

\keyword{datasets}
