% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_state.R
\docType{data}
\name{StepState}
\alias{StepState}
\title{StepState}
\format{\code{\link{R6Class}} object.}
\usage{
StepState
}
\description{
StepState
}
\section{Fields}{

\describe{
\item{\code{taskId}}{of type String.}

\item{\code{taskState}}{object of class \code{\link{State}}.}
}}

\keyword{datasets}
