% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_profile.R
\docType{data}
\name{StorageProfile}
\alias{StorageProfile}
\title{StorageProfile}
\format{\code{\link{R6Class}} object.}
\usage{
StorageProfile
}
\description{
StorageProfile
}
\section{Fields}{

\describe{
\item{\code{name}}{of type String.}

\item{\code{size}}{of type int.}
}}

\keyword{datasets}
