% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_color_element.R
\docType{data}
\name{StringColorElement}
\alias{StringColorElement}
\title{StringColorElement}
\format{\code{\link{R6Class}} object, super class \code{\link{ColorElement}}.}
\usage{
StringColorElement
}
\description{
StringColorElement
}
\section{Fields}{

\describe{
\item{\code{color}}{of type int inherited from super class \code{\link{ColorElement}}.}

\item{\code{stringValue}}{of type String.}
}}

\keyword{datasets}
