% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscription_plan_service.R
\docType{data}
\name{SubscriptionPlanService}
\alias{SubscriptionPlanService}
\title{SubscriptionPlanService}
\format{\code{\link{R6Class}} object.}
\usage{
SubscriptionPlanService
}
\description{
SubscriptionPlanService
}
\section{Methods}{

\describe{
   \item{\code{getSubscriptionPlans(userId)}}{method}
   \item{\code{getPlans(userId)}}{method}
   \item{\code{createSubscriptionPlan(userId,plan,successUrl,cancelUrl)}}{method}
   \item{\code{setSubscriptionPlanStatus(subscriptionPlanId,status)}}{method}
   \item{\code{updatePaymentMethod(subscriptionPlanId,successUrl,cancelUrl)}}{method}
   \item{\code{setUpdatePaymentMethodStatus(subscriptionPlanId,status)}}{method}
   \item{\code{cancelSubscription(subscriptionPlanId)}}{method}
   \item{\code{upgradeSubscription(subscriptionPlanId,plan)}}{method}
}
}

\keyword{datasets}
