% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\docType{data}
\name{Table}
\alias{Table}
\title{Table}
\format{\code{\link{R6Class}} object.}
\usage{
Table
}
\description{
Table
}
\section{Fields}{

\describe{
\item{\code{nRows}}{of type int.}

\item{\code{properties}}{object of class \code{\link{TableProperties}}.}

\item{\code{columns}}{list of class \code{\link{Column}}.}
}}

\keyword{datasets}
