% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_profile.R
\docType{data}
\name{TableProfile}
\alias{TableProfile}
\title{TableProfile}
\format{\code{\link{R6Class}} object, super class \code{\link{Profile}}.}
\usage{
TableProfile
}
\description{
TableProfile
}
\section{Fields}{

\describe{
\item{\code{name}}{of type String inherited from super class \code{\link{Profile}}.}

\item{\code{nRows}}{of type int.}

\item{\code{nCols}}{of type int.}
}}

\keyword{datasets}
