% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_properties.R
\docType{data}
\name{TableProperties}
\alias{TableProperties}
\title{TableProperties}
\format{\code{\link{R6Class}} object.}
\usage{
TableProperties
}
\description{
TableProperties
}
\section{Fields}{

\describe{
\item{\code{name}}{of type String.}

\item{\code{sortOrder}}{list of type String.}

\item{\code{ascending}}{of type bool.}
}}

\keyword{datasets}
