% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_schema_service.R
\docType{data}
\name{TableSchemaService}
\alias{TableSchemaService}
\title{TableSchemaService}
\format{\code{\link{R6Class}} object.}
\usage{
TableSchemaService
}
\description{
TableSchemaService
}
\section{Methods}{

\describe{
   \item{\code{select(tableId,cnames,offset,limit)}}{method}
   \item{\code{selectPairwise(tableId,cnames,offset,limit)}}{method}
   \item{\code{selectStream(tableId,cnames,offset,limit)}}{method}
   \item{\code{selectCSV(tableId,cnames,offset,limit,separator,quote,encoding)}}{method}
}
}

\keyword{datasets}
