% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_step_model.R
\docType{data}
\name{TableStepModel}
\alias{TableStepModel}
\title{TableStepModel}
\format{\code{\link{R6Class}} object, super class \code{\link{StepModel}}.}
\usage{
TableStepModel
}
\description{
TableStepModel
}
\section{Fields}{

\describe{
\item{\code{relation}}{object of class \code{\link{Relation}}.}
}}

\keyword{datasets}
