% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_data_event.R
\docType{data}
\name{TaskDataEvent}
\alias{TaskDataEvent}
\title{TaskDataEvent}
\format{\code{\link{R6Class}} object, super class \code{\link{TaskEvent}}.}
\usage{
TaskDataEvent
}
\description{
TaskDataEvent
}
\section{Fields}{

\describe{
\item{\code{taskId}}{of type String inherited from super class \code{\link{TaskEvent}}.}

\item{\code{isDeleted}}{of type bool inherited from super class \code{\link{PersistentObject}}.}

\item{\code{rev}}{of type String inherited from super class \code{\link{PersistentObject}}.}

\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{bytes}}{of type Uint8List.}

\item{\code{date}}{object of class \code{\link{Date}} inherited from super class \code{\link{Event}}.}
}}

\keyword{datasets}
