% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_event.R
\docType{data}
\name{TaskEvent}
\alias{TaskEvent}
\title{TaskEvent}
\format{\code{\link{R6Class}} object, super class \code{\link{Event}}, sub classes \code{\link{TaskLogEvent}}, \code{\link{TaskProgressEvent}}, \code{\link{TaskDataEvent}}, \code{\link{TaskStateEvent}}.}
\usage{
TaskEvent
}
\description{
TaskEvent
}
\section{Fields}{

\describe{
\item{\code{isDeleted}}{of type bool inherited from super class \code{\link{PersistentObject}}.}

\item{\code{rev}}{of type String inherited from super class \code{\link{PersistentObject}}.}

\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{taskId}}{of type String.}

\item{\code{date}}{object of class \code{\link{Date}} inherited from super class \code{\link{Event}}.}
}}

\keyword{datasets}
