% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_progress_event.R
\docType{data}
\name{TaskProgressEvent}
\alias{TaskProgressEvent}
\title{TaskProgressEvent}
\format{\code{\link{R6Class}} object, super class \code{\link{TaskEvent}}.}
\usage{
TaskProgressEvent
}
\description{
TaskProgressEvent
}
\section{Fields}{

\describe{
\item{\code{taskId}}{of type String inherited from super class \code{\link{TaskEvent}}.}

\item{\code{isDeleted}}{of type bool inherited from super class \code{\link{PersistentObject}}.}

\item{\code{rev}}{of type String inherited from super class \code{\link{PersistentObject}}.}

\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{total}}{of type int.}

\item{\code{actual}}{of type int.}

\item{\code{date}}{object of class \code{\link{Date}} inherited from super class \code{\link{Event}}.}
}}

\keyword{datasets}
