% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_resource.R
\docType{data}
\name{TaskResource}
\alias{TaskResource}
\title{TaskResource}
\format{\code{\link{R6Class}} object, sub classes \code{\link{SimpleTaskResource}}, \code{\link{WorkerTaskResource}}.}
\usage{
TaskResource
}
\description{
TaskResource
}
\keyword{datasets}
