% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_service.R
\docType{data}
\name{TaskService}
\alias{TaskService}
\title{TaskService}
\format{\code{\link{R6Class}} object.}
\usage{
TaskService
}
\description{
TaskService
}
\section{Methods}{

\describe{
   \item{\code{runTask(taskId)}}{method}
   \item{\code{cancelTask(taskId)}}{method}
   \item{\code{waitDone(taskId)}}{method}
   \item{\code{updateWorker(worker)}}{method}
   \item{\code{taskDurationByTeam(teamId,year,month)}}{method}
}
}

\keyword{datasets}
