% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_summary.R
\docType{data}
\name{TaskSummary}
\alias{TaskSummary}
\title{TaskSummary}
\format{\code{\link{R6Class}} object.}
\usage{
TaskSummary
}
\description{
TaskSummary
}
\section{Fields}{

\describe{
\item{\code{n}}{of type int.}

\item{\code{duration}}{of type double.}
}}

\keyword{datasets}
