% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_id.R
\docType{data}
\name{TaxId}
\alias{TaxId}
\title{TaxId}
\format{\code{\link{R6Class}} object.}
\usage{
TaxId
}
\description{
TaxId
}
\section{Fields}{

\describe{
\item{\code{type}}{of type String.}

\item{\code{value}}{of type String.}

\item{\code{isValid}}{of type bool.}
}}

\keyword{datasets}
