% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_service.R
\docType{data}
\name{TeamService}
\alias{TeamService}
\title{TeamService}
\format{\code{\link{R6Class}} object.}
\usage{
TeamService
}
\description{
TeamService
}
\section{Methods}{

\describe{
   \item{\code{profiles(teamId)}}{method}
   \item{\code{resourceSummary(teamId)}}{method}
}
}

\keyword{datasets}
