% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\docType{data}
\name{Token}
\alias{Token}
\title{Token}
\format{\code{\link{R6Class}} object.}
\usage{
Token
}
\description{
Token
}
\section{Fields}{

\describe{
\item{\code{userId}}{of type String.}

\item{\code{token}}{of type String.}

\item{\code{issuedAt}}{object of class \code{\link{Date}}.}

\item{\code{expiry}}{object of class \code{\link{Date}}.}
}}

\keyword{datasets}
