% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulimits.R
\docType{data}
\name{Ulimits}
\alias{Ulimits}
\title{Ulimits}
\format{\code{\link{R6Class}} object.}
\usage{
Ulimits
}
\description{
Ulimits
}
\section{Fields}{

\describe{
\item{\code{core_file_size}}{of type int.}

\item{\code{data_seg_size}}{of type int.}

\item{\code{scheduling_priority}}{of type int.}

\item{\code{file_size}}{of type int.}

\item{\code{pending_signals}}{of type int.}

\item{\code{max_locked_memory}}{of type int.}

\item{\code{max_memory_size}}{of type int.}

\item{\code{open_files}}{of type int.}

\item{\code{pipe_size}}{of type int.}

\item{\code{message_queues}}{of type int.}

\item{\code{real_time_priority}}{of type int.}

\item{\code{stack_size}}{of type int.}

\item{\code{cpu_time}}{of type int.}

\item{\code{max_user_processes}}{of type int.}

\item{\code{virtual_memory}}{of type int.}

\item{\code{file_locks}}{of type int.}
}}

\keyword{datasets}
