% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/union_relation.R
\docType{data}
\name{UnionRelation}
\alias{UnionRelation}
\title{UnionRelation}
\format{\code{\link{R6Class}} object, super class \code{\link{Relation}}.}
\usage{
UnionRelation
}
\description{
UnionRelation
}
\section{Fields}{

\describe{
\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{relations}}{list of class \code{\link{Relation}}.}
}}

\keyword{datasets}
