% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_model.R
\docType{data}
\name{UrlModel}
\alias{UrlModel}
\title{UrlModel}
\format{\code{\link{R6Class}} object, super class \code{\link{StepModel}}.}
\usage{
UrlModel
}
\description{
UrlModel
}
\section{Fields}{

\describe{
\item{\code{url}}{of type String.}
}}

\keyword{datasets}
