% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\docType{data}
\name{User}
\alias{User}
\title{User}
\format{\code{\link{R6Class}} object, super class \code{\link{Document}}, sub classes \code{\link{Team}}.}
\usage{
User
}
\description{
User
}
\section{Fields}{

\describe{
\item{\code{description}}{of type String inherited from super class \code{\link{Document}}.}

\item{\code{name}}{of type String inherited from super class \code{\link{Document}}.}

\item{\code{createdBy}}{of type String inherited from super class \code{\link{Document}}.}

\item{\code{tags}}{list of type String inherited from super class \code{\link{Document}}.}

\item{\code{version}}{of type String inherited from super class \code{\link{Document}}.}

\item{\code{authors}}{list of type String inherited from super class \code{\link{Document}}.}

\item{\code{isPublic}}{of type bool inherited from super class \code{\link{Document}}.}

\item{\code{isDeleted}}{of type bool inherited from super class \code{\link{PersistentObject}}.}

\item{\code{rev}}{of type String inherited from super class \code{\link{PersistentObject}}.}

\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{email}}{of type String.}

\item{\code{isValidated}}{of type bool.}

\item{\code{domain}}{of type String.}

\item{\code{roles}}{list of type String.}

\item{\code{invitedByUsername}}{of type String.}

\item{\code{invitationCounts}}{of type int.}

\item{\code{maxInvitation}}{of type int.}

\item{\code{acl}}{object of class \code{\link{Acl}} inherited from super class \code{\link{Document}}.}

\item{\code{createdDate}}{object of class \code{\link{Date}} inherited from super class \code{\link{Document}}.}

\item{\code{lastModifiedDate}}{object of class \code{\link{Date}} inherited from super class \code{\link{Document}}.}

\item{\code{urls}}{list of class \code{\link{Url}} inherited from super class \code{\link{Document}}.}

\item{\code{meta}}{list of class \code{\link{Pair}} inherited from super class \code{\link{Document}}.}

\item{\code{url}}{object of class \code{\link{Url}} inherited from super class \code{\link{Document}}.}

\item{\code{teamAcl}}{object of class \code{\link{Acl}}.}

\item{\code{billingInfo}}{object of class \code{\link{BillingInfo}}.}
}}

\keyword{datasets}
