% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_secret.R
\docType{data}
\name{UserSecret}
\alias{UserSecret}
\title{UserSecret}
\format{\code{\link{R6Class}} object, super class \code{\link{PersistentObject}}.}
\usage{
UserSecret
}
\description{
UserSecret
}
\section{Fields}{

\describe{
\item{\code{isDeleted}}{of type bool inherited from super class \code{\link{PersistentObject}}.}

\item{\code{rev}}{of type String inherited from super class \code{\link{PersistentObject}}.}

\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{userId}}{of type String.}

\item{\code{salt}}{of type String.}

\item{\code{hashPassword}}{of type String.}
}}

\keyword{datasets}
