% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_secret_service.R
\docType{data}
\name{UserSecretService}
\alias{UserSecretService}
\title{UserSecretService}
\format{\code{\link{R6Class}} object.}
\usage{
UserSecretService
}
\description{
UserSecretService
}
\section{Methods}{

\describe{
}
}

\keyword{datasets}
