% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_service.R
\docType{data}
\name{UserService}
\alias{UserService}
\title{UserService}
\format{\code{\link{R6Class}} object.}
\usage{
UserService
}
\description{
UserService
}
\section{Methods}{

\describe{
   \item{\code{cookieConsent(dummy)}}{method}
   \item{\code{connect(usernameOrEmail,password)}}{method}
   \item{\code{connect2(domain,usernameOrEmail,password)}}{method}
   \item{\code{createUser(user,password)}}{method}
   \item{\code{hasUserName(username)}}{method}
   \item{\code{updatePassword(userId,password)}}{method}
   \item{\code{updateBillingInfo(userId,billingInfo)}}{method}
   \item{\code{viesInfo(country_code,vatNumber)}}{method}
   \item{\code{summary(userId)}}{method}
   \item{\code{resourceSummary(userId)}}{method}
   \item{\code{profiles(userId)}}{method}
   \item{\code{createToken(userId,validityInSeconds)}}{method}
   \item{\code{isTokenValid(token)}}{method}
   \item{\code{setTeamPrivilege(username,principal,privilege)}}{method}
   \item{\code{getServerVersion(module)}}{method}
   \item{\code{getInvited(email)}}{method}
   \item{\code{sendValidationMail(email)}}{method}
   \item{\code{sendResetPasswordEmail(email)}}{method}
   \item{\code{changeUserPassword(token,password)}}{method}
   \item{\code{validateUser(token)}}{method}
   \item{\code{canCreatePrivateProject(teamOrUserId)}}{method}
}
}

\keyword{datasets}
