% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_session.R
\docType{data}
\name{UserSession}
\alias{UserSession}
\title{UserSession}
\format{\code{\link{R6Class}} object.}
\usage{
UserSession
}
\description{
UserSession
}
\section{Fields}{

\describe{
\item{\code{serverVersion}}{object of class \code{\link{Version}}.}

\item{\code{user}}{object of class \code{\link{User}}.}

\item{\code{token}}{object of class \code{\link{Token}}.}
}}

\keyword{datasets}
