% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\docType{data}
\name{Version}
\alias{Version}
\title{Version}
\format{\code{\link{R6Class}} object.}
\usage{
Version
}
\description{
Version
}
\section{Fields}{

\describe{
\item{\code{major}}{of type int.}

\item{\code{minor}}{of type int.}

\item{\code{patch}}{of type int.}

\item{\code{tag}}{of type String.}

\item{\code{date}}{of type String.}
}}

\keyword{datasets}
