% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vies_info.R
\docType{data}
\name{ViesInfo}
\alias{ViesInfo}
\title{ViesInfo}
\format{\code{\link{R6Class}} object.}
\usage{
ViesInfo
}
\description{
ViesInfo
}
\section{Fields}{

\describe{
\item{\code{countryCode}}{of type String.}

\item{\code{vatNumber}}{of type String.}

\item{\code{requestDate}}{of type String.}

\item{\code{valid}}{of type bool.}

\item{\code{name}}{of type String.}

\item{\code{address}}{of type String.}
}}

\keyword{datasets}
