% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_data_step.R
\docType{data}
\name{ViewDataStep}
\alias{ViewDataStep}
\title{ViewDataStep}
\format{\code{\link{R6Class}} object, super class \code{\link{CrossTabStep}}.}
\usage{
ViewDataStep
}
\description{
ViewDataStep
}
\section{Fields}{

\describe{
\item{\code{groupId}}{of type String inherited from super class \code{\link{Step}}.}

\item{\code{name}}{of type String inherited from super class \code{\link{Step}}.}

\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{model}}{object of class \code{\link{Crosstab}} inherited from super class \code{\link{CrossTabStep}}.}

\item{\code{model}}{object of class \code{\link{StepModel}} inherited from super class \code{\link{ModelStep}}.}

\item{\code{inputs}}{list of class \code{\link{InputPort}} inherited from super class \code{\link{Step}}.}

\item{\code{outputs}}{list of class \code{\link{OutputPort}} inherited from super class \code{\link{Step}}.}

\item{\code{rectangle}}{object of class \code{\link{Rectangle}} inherited from super class \code{\link{Step}}.}

\item{\code{state}}{object of class \code{\link{StepState}} inherited from super class \code{\link{Step}}.}

\item{\code{model}}{object of class \code{\link{StepModel}} inherited from super class \code{\link{Step}}.}
}}

\keyword{datasets}
