% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/where_relation.R
\docType{data}
\name{WhereRelation}
\alias{WhereRelation}
\title{WhereRelation}
\format{\code{\link{R6Class}} object, super class \code{\link{Relation}}.}
\usage{
WhereRelation
}
\description{
WhereRelation
}
\section{Fields}{

\describe{
\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{relation}}{object of class \code{\link{Relation}}.}

\item{\code{filters}}{object of class \code{\link{Filters}}.}
}}

\keyword{datasets}
