% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizard_step_model.R
\docType{data}
\name{WizardStepModel}
\alias{WizardStepModel}
\title{WizardStepModel}
\format{\code{\link{R6Class}} object, super class \code{\link{StepModel}}.}
\usage{
WizardStepModel
}
\description{
WizardStepModel
}
\section{Fields}{

\describe{
\item{\code{name}}{of type String.}

\item{\code{description}}{of type String.}

\item{\code{factors}}{list of class \code{\link{MappingFactor}}.}

\item{\code{steps}}{list of class \code{\link{Step}}.}
}}

\keyword{datasets}
