% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker_endpoint_running_environment.R
\docType{data}
\name{WorkerEndpointRunningEnvironment}
\alias{WorkerEndpointRunningEnvironment}
\title{WorkerEndpointRunningEnvironment}
\format{\code{\link{R6Class}} object, super class \code{\link{RunningEnvironment}}.}
\usage{
WorkerEndpointRunningEnvironment
}
\description{
WorkerEndpointRunningEnvironment
}
\section{Fields}{

\describe{
\item{\code{workerMachineId}}{of type String.}
}}

\keyword{datasets}
