% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker_endpoint_service.R
\docType{data}
\name{WorkerEndpointService}
\alias{WorkerEndpointService}
\title{WorkerEndpointService}
\format{\code{\link{R6Class}} object.}
\usage{
WorkerEndpointService
}
\description{
WorkerEndpointService
}
\section{Methods}{

\describe{
}
}

\keyword{datasets}
