% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker_running_environment.R
\docType{data}
\name{WorkerRunningEnvironment}
\alias{WorkerRunningEnvironment}
\title{WorkerRunningEnvironment}
\format{\code{\link{R6Class}} object, super class \code{\link{RunningEnvironment}}.}
\usage{
WorkerRunningEnvironment
}
\description{
WorkerRunningEnvironment
}
\section{Fields}{

\describe{
\item{\code{workerMachineId}}{of type String.}
}}

\keyword{datasets}
