% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker_service.R
\docType{data}
\name{WorkerService}
\alias{WorkerService}
\title{WorkerService}
\format{\code{\link{R6Class}} object.}
\usage{
WorkerService
}
\description{
WorkerService
}
\section{Methods}{

\describe{
   \item{\code{exec(task)}}{method}
   \item{\code{setPriority(priority)}}{method}
   \item{\code{setStatus(status)}}{method}
   \item{\code{setHeartBeat(heartBeat)}}{method}
   \item{\code{getState(all)}}{method}
}
}

\keyword{datasets}
