% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker_task_resource.R
\docType{data}
\name{WorkerTaskResource}
\alias{WorkerTaskResource}
\title{WorkerTaskResource}
\format{\code{\link{R6Class}} object, super class \code{\link{RunningEnvironment}}.}
\usage{
WorkerTaskResource
}
\description{
WorkerTaskResource
}
\section{Fields}{

\describe{
\item{\code{workerMachineId}}{of type String.}
}}

\keyword{datasets}
