% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\docType{data}
\name{Workflow}
\alias{Workflow}
\title{Workflow}
\format{\code{\link{R6Class}} object, super class \code{\link{ProjectDocument}}.}
\usage{
Workflow
}
\description{
Workflow
}
\section{Fields}{

\describe{
\item{\code{projectId}}{of type String inherited from super class \code{\link{ProjectDocument}}.}

\item{\code{description}}{of type String inherited from super class \code{\link{Document}}.}

\item{\code{name}}{of type String inherited from super class \code{\link{Document}}.}

\item{\code{createdBy}}{of type String inherited from super class \code{\link{Document}}.}

\item{\code{tags}}{list of type String inherited from super class \code{\link{Document}}.}

\item{\code{isDeleted}}{of type bool inherited from super class \code{\link{PersistentObject}}.}

\item{\code{rev}}{of type String inherited from super class \code{\link{PersistentObject}}.}

\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{isApp}}{of type bool.}

\item{\code{isTemplate}}{of type bool.}

\item{\code{acl}}{object of class \code{\link{Acl}} inherited from super class \code{\link{Document}}.}

\item{\code{createdDate}}{object of class \code{\link{Date}} inherited from super class \code{\link{Document}}.}

\item{\code{lastModifiedDate}}{object of class \code{\link{Date}} inherited from super class \code{\link{Document}}.}

\item{\code{urls}}{list of class \code{\link{Url}} inherited from super class \code{\link{Document}}.}

\item{\code{links}}{list of class \code{\link{Link}}.}

\item{\code{steps}}{list of class \code{\link{Step}}.}

\item{\code{offset}}{object of class \code{\link{Point}}.}
}}

\keyword{datasets}
