% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_document.R
\docType{data}
\name{WorkflowDocument}
\alias{WorkflowDocument}
\title{WorkflowDocument}
\format{\code{\link{R6Class}} object, super class \code{\link{FileDocument}}.}
\usage{
WorkflowDocument
}
\description{
WorkflowDocument
}
\section{Fields}{

\describe{
\item{\code{dataUri}}{of type String inherited from super class \code{\link{FileDocument}}.}

\item{\code{size}}{of type int inherited from super class \code{\link{FileDocument}}.}

\item{\code{projectId}}{of type String inherited from super class \code{\link{ProjectDocument}}.}

\item{\code{description}}{of type String inherited from super class \code{\link{Document}}.}

\item{\code{name}}{of type String inherited from super class \code{\link{Document}}.}

\item{\code{createdBy}}{of type String inherited from super class \code{\link{Document}}.}

\item{\code{tags}}{list of type String inherited from super class \code{\link{Document}}.}

\item{\code{version}}{of type String inherited from super class \code{\link{Document}}.}

\item{\code{authors}}{list of type String inherited from super class \code{\link{Document}}.}

\item{\code{isDeleted}}{of type bool inherited from super class \code{\link{PersistentObject}}.}

\item{\code{rev}}{of type String inherited from super class \code{\link{PersistentObject}}.}

\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{workflowId}}{of type String.}

\item{\code{stepId}}{of type String.}

\item{\code{metadata}}{object of class \code{\link{FileMetadata}} inherited from super class \code{\link{FileDocument}}.}

\item{\code{acl}}{object of class \code{\link{Acl}} inherited from super class \code{\link{Document}}.}

\item{\code{createdDate}}{object of class \code{\link{Date}} inherited from super class \code{\link{Document}}.}

\item{\code{lastModifiedDate}}{object of class \code{\link{Date}} inherited from super class \code{\link{Document}}.}

\item{\code{urls}}{list of class \code{\link{Url}} inherited from super class \code{\link{Document}}.}

\item{\code{meta}}{list of class \code{\link{Pair}} inherited from super class \code{\link{Document}}.}

\item{\code{url}}{object of class \code{\link{Url}} inherited from super class \code{\link{Document}}.}
}}

\keyword{datasets}
