% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_service.R
\docType{data}
\name{WorkflowService}
\alias{WorkflowService}
\title{WorkflowService}
\format{\code{\link{R6Class}} object.}
\usage{
WorkflowService
}
\description{
WorkflowService
}
\section{Methods}{

\describe{
   \item{\code{getCubeQuery(workflowId,stepId)}}{method}
   \item{\code{copy(workflowId)}}{method}
   \item{\code{cloneWorkflow(workflowId,teamId,projectId)}}{method}
}
}

\keyword{datasets}
