% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_y_axis.R
\docType{data}
\name{XYAxis}
\alias{XYAxis}
\title{XYAxis}
\format{\code{\link{R6Class}} object.}
\usage{
XYAxis
}
\description{
XYAxis
}
\section{Fields}{

\describe{
\item{\code{taskId}}{of type String.}

\item{\code{chart}}{object of class \code{\link{Chart}}.}

\item{\code{colors}}{object of class \code{\link{Colors}}.}

\item{\code{errors}}{object of class \code{\link{Errors}}.}

\item{\code{labels}}{object of class \code{\link{Labels}}.}

\item{\code{xAxis}}{object of class \code{\link{Axis}}.}

\item{\code{yAxis}}{object of class \code{\link{Axis}}.}
}}

\keyword{datasets}
