% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_y_axis_list.R
\docType{data}
\name{XYAxisList}
\alias{XYAxisList}
\title{XYAxisList}
\format{\code{\link{R6Class}} object.}
\usage{
XYAxisList
}
\description{
XYAxisList
}
\section{Fields}{

\describe{
\item{\code{rectangleSelections}}{list of class \code{\link{Rectangle}}.}

\item{\code{xyAxis}}{list of class \code{\link{XYAxis}}.}
}}

\keyword{datasets}
