% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acl.R
\docType{data}
\name{Acl}
\alias{Acl}
\title{Acl}
\format{\code{\link{R6Class}} object.}
\usage{
Acl
}
\description{
Acl
}
\section{Fields}{

\describe{
\item{\code{owner}}{of type String.}

\item{\code{resourceId}}{of type String.}

\item{\code{resourceType}}{of type String.}

\item{\code{aces}}{list of class \code{\link{Ace}}.}
}}

\keyword{datasets}
