% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acl_context.R
\docType{data}
\name{AclContext}
\alias{AclContext}
\title{AclContext}
\format{\code{\link{R6Class}} object.}
\usage{
AclContext
}
\description{
AclContext
}
\section{Fields}{

\describe{
\item{\code{username}}{of type String.}
}}

\keyword{datasets}
