% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.R
\docType{data}
\name{Chart}
\alias{Chart}
\title{Chart}
\format{\code{\link{R6Class}} object, sub classes \code{\link{ChartLine}}, \code{\link{ChartPoint}}, \code{\link{ChartHeatmap}}, \code{\link{ChartSize}}, \code{\link{ChartBar}}.}
\usage{
Chart
}
\description{
Chart
}
\keyword{datasets}
