% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_heatmap.R
\docType{data}
\name{ChartHeatmap}
\alias{ChartHeatmap}
\title{ChartHeatmap}
\format{\code{\link{R6Class}} object, super class \code{\link{Chart}}.}
\usage{
ChartHeatmap
}
\description{
ChartHeatmap
}
\section{Fields}{

\describe{
\item{\code{name}}{of type String inherited from super class \code{\link{Chart}}.}

\item{\code{properties}}{object of class \code{\link{Properties}} inherited from super class \code{\link{Chart}}.}
}}

\keyword{datasets}
