% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_list.R
\docType{data}
\name{ColorList}
\alias{ColorList}
\title{ColorList}
\format{\code{\link{R6Class}} object.}
\usage{
ColorList
}
\description{
ColorList
}
\section{Fields}{

\describe{
\item{\code{name}}{of type String.}
}}

\keyword{datasets}
