% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_schema_meta_data.R
\docType{data}
\name{ColumnSchemaMetaData}
\alias{ColumnSchemaMetaData}
\title{ColumnSchemaMetaData}
\format{\code{\link{R6Class}} object.}
\usage{
ColumnSchemaMetaData
}
\description{
ColumnSchemaMetaData
}
\section{Fields}{

\describe{
\item{\code{sort}}{list of type String.}

\item{\code{ascending}}{of type bool.}

\item{\code{quartiles}}{list of type String.}

\item{\code{properties}}{list of class \code{\link{Pair}}.}
}}

\keyword{datasets}
