% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_library_service.R
\docType{data}
\name{CranLibraryService}
\alias{CranLibraryService}
\title{CranLibraryService}
\format{\code{\link{R6Class}} object.}
\usage{
CranLibraryService
}
\description{
CranLibraryService
}
\section{Methods}{

\describe{
   \item{\code{packagesGz(repoName)}}{method}
   \item{\code{packagesRds(repoName)}}{method}
   \item{\code{packages(repoName)}}{method}
   \item{\code{archive(repoName,package,filename)}}{method}
   \item{\code{package(repoName,package)}}{method}
}
}

\keyword{datasets}
