% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\docType{data}
\name{Date}
\alias{Date}
\title{Date}
\format{\code{\link{R6Class}} object.}
\usage{
Date
}
\description{
Date
}
\section{Fields}{

\describe{
\item{\code{value}}{of type String.}
}}

\keyword{datasets}
