% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_service.R
\name{DocumentService}
\alias{DocumentService}
\title{DocumentService}
\format{\code{\link{R6Class}} object.}
\description{
DocumentService

DocumentService
}
\section{Methods}{

\describe{
   \item{\code{getTercenOperatorLibrary(offset,limit)}}{method}
   \item{\code{getTercenWorkflowLibrary(offset,limit)}}{method}
   \item{\code{getTercenAppLibrary(offset,limit)}}{method}
}
}

\section{Super class}{
\code{\link[tercen:HttpClientService]{tercen::HttpClientService}} -> \code{DocumentService}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DocumentService$new()}}
\item \href{#method-findProjectByOwnersAndName}{\code{DocumentService$findProjectByOwnersAndName()}}
\item \href{#method-findProjectByOwnersAndCreatedDate}{\code{DocumentService$findProjectByOwnersAndCreatedDate()}}
\item \href{#method-findWorkflowByTagOwnerCreatedDate}{\code{DocumentService$findWorkflowByTagOwnerCreatedDate()}}
\item \href{#method-findOperatorByOwnerLastModifiedDate}{\code{DocumentService$findOperatorByOwnerLastModifiedDate()}}
\item \href{#method-findOperatorByCreatedDate}{\code{DocumentService$findOperatorByCreatedDate()}}
\item \href{#method-getTercenOperatorLibrary}{\code{DocumentService$getTercenOperatorLibrary()}}
\item \href{#method-getTercenWorkflowLibrary}{\code{DocumentService$getTercenWorkflowLibrary()}}
\item \href{#method-getTercenAppLibrary}{\code{DocumentService$getTercenAppLibrary()}}
\item \href{#method-clone}{\code{DocumentService$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="create">}\href{../../tercen/html/HttpClientService.html#method-create}{\code{tercen::HttpClientService$create()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="delete">}\href{../../tercen/html/HttpClientService.html#method-delete}{\code{tercen::HttpClientService$delete()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="findKeys">}\href{../../tercen/html/HttpClientService.html#method-findKeys}{\code{tercen::HttpClientService$findKeys()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="findStartKeys">}\href{../../tercen/html/HttpClientService.html#method-findStartKeys}{\code{tercen::HttpClientService$findStartKeys()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="fromTson">}\href{../../tercen/html/HttpClientService.html#method-fromTson}{\code{tercen::HttpClientService$fromTson()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="get">}\href{../../tercen/html/HttpClientService.html#method-get}{\code{tercen::HttpClientService$get()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="getServiceUri">}\href{../../tercen/html/HttpClientService.html#method-getServiceUri}{\code{tercen::HttpClientService$getServiceUri()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="list">}\href{../../tercen/html/HttpClientService.html#method-list}{\code{tercen::HttpClientService$list()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="onResponseError">}\href{../../tercen/html/HttpClientService.html#method-onResponseError}{\code{tercen::HttpClientService$onResponseError()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="toTson">}\href{../../tercen/html/HttpClientService.html#method-toTson}{\code{tercen::HttpClientService$toTson()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tercen" data-topic="HttpClientService" data-id="update">}\href{../../tercen/html/HttpClientService.html#method-update}{\code{tercen::HttpClientService$update()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DocumentService$new(baseRestUri, client)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-findProjectByOwnersAndName"></a>}}
\subsection{Method \code{findProjectByOwnersAndName()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DocumentService$findProjectByOwnersAndName(
  startKey = NULL,
  endKey = NULL,
  limit = 20,
  skip = 0,
  descending = TRUE,
  useFactory = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-findProjectByOwnersAndCreatedDate"></a>}}
\subsection{Method \code{findProjectByOwnersAndCreatedDate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DocumentService$findProjectByOwnersAndCreatedDate(
  startKey = NULL,
  endKey = NULL,
  limit = 20,
  skip = 0,
  descending = TRUE,
  useFactory = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-findWorkflowByTagOwnerCreatedDate"></a>}}
\subsection{Method \code{findWorkflowByTagOwnerCreatedDate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DocumentService$findWorkflowByTagOwnerCreatedDate(
  startKey = NULL,
  endKey = NULL,
  limit = 20,
  skip = 0,
  descending = TRUE,
  useFactory = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-findOperatorByOwnerLastModifiedDate"></a>}}
\subsection{Method \code{findOperatorByOwnerLastModifiedDate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DocumentService$findOperatorByOwnerLastModifiedDate(
  startKey = NULL,
  endKey = NULL,
  limit = 20,
  skip = 0,
  descending = TRUE,
  useFactory = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-findOperatorByCreatedDate"></a>}}
\subsection{Method \code{findOperatorByCreatedDate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DocumentService$findOperatorByCreatedDate(
  startKey = NULL,
  endKey = NULL,
  limit = 20,
  skip = 0,
  descending = TRUE,
  useFactory = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getTercenOperatorLibrary"></a>}}
\subsection{Method \code{getTercenOperatorLibrary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DocumentService$getTercenOperatorLibrary(offset, limit)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getTercenWorkflowLibrary"></a>}}
\subsection{Method \code{getTercenWorkflowLibrary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DocumentService$getTercenWorkflowLibrary(offset, limit)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getTercenAppLibrary"></a>}}
\subsection{Method \code{getTercenAppLibrary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DocumentService$getTercenAppLibrary(offset, limit)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DocumentService$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
