% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\docType{data}
\name{Errors}
\alias{Errors}
\title{Errors}
\format{\code{\link{R6Class}} object.}
\usage{
Errors
}
\description{
Errors
}
\section{Fields}{

\describe{
\item{\code{factors}}{list of class \code{\link{Factor}}.}
}}

\keyword{datasets}
