% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by_relation.R
\docType{data}
\name{GroupByRelation}
\alias{GroupByRelation}
\title{GroupByRelation}
\format{\code{\link{R6Class}} object, super class \code{\link{Relation}}.}
\usage{
GroupByRelation
}
\description{
GroupByRelation
}
\section{Fields}{

\describe{
\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{group}}{list of type String.}

\item{\code{relation}}{object of class \code{\link{Relation}}.}
}}

\keyword{datasets}
