% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_object.R
\docType{data}
\name{IdObject}
\alias{IdObject}
\title{IdObject}
\format{\code{\link{R6Class}} object, sub classes \code{\link{DataStep}}, \code{\link{MeltStep}}, \code{\link{JoinStep}}, \code{\link{WizardStep}}, \code{\link{CrossTabStep}}, \code{\link{InStep}}, \code{\link{GroupStep}}, \code{\link{OutStep}}, \code{\link{TableStep}}, \code{\link{NamespaceStep}}, \code{\link{RelationStep}}, \code{\link{ExportStep}}, \code{\link{ModelStep}}, \code{\link{ViewStep}}, \code{\link{InputPort}}, \code{\link{OutputPort}}, \code{\link{GarbageTasks}}, \code{\link{ComputationTask}}, \code{\link{ExportWorkflowTask}}, \code{\link{RunWebAppTask}}, \code{\link{CSVTask}}, \code{\link{CubeQueryTask}}, \code{\link{ExportTableTask}}, \code{\link{ImportWorkflowTask}}, \code{\link{ImportGitWorkflowTask}}, \code{\link{GlTask}}, \code{\link{ProjectTask}}, \code{\link{CreateGitOperatorTask}}, \code{\link{TaskStateEvent}}, \code{\link{TaskLogEvent}}, \code{\link{TaskProgressEvent}}, \code{\link{TaskDataEvent}}, \code{\link{TaskEvent}}, \code{\link{Team}}, \code{\link{ShinyOperator}}, \code{\link{ROperator}}, \code{\link{WebAppOperator}}, \code{\link{GitOperator}}, \code{\link{CubeQueryTableSchema}}, \code{\link{TableSchema}}, \code{\link{ComputedTableSchema}}, \code{\link{Schema}}, \code{\link{FileDocument}}, \code{\link{Workflow}}, \code{\link{User}}, \code{\link{Operator}}, \code{\link{ProjectDocument}}, \code{\link{Project}}, \code{\link{GarbageObject}}, \code{\link{Task}}, \code{\link{Event}}, \code{\link{Document}}, \code{\link{Lock}}, \code{\link{UserSecret}}, \code{\link{WhereRelation}}, \code{\link{RenameRelation}}, \code{\link{DistinctRelation}}, \code{\link{UnionRelation}}, \code{\link{SimpleRelation}}, \code{\link{InMemoryRelation}}, \code{\link{GatherRelation}}, \code{\link{CompositeRelation}}, \code{\link{GroupByRelation}}, \code{\link{Column}}, \code{\link{StartProcess}}, \code{\link{Step}}, \code{\link{Port}}, \code{\link{PersistentObject}}, \code{\link{Link}}, \code{\link{Relation}}, \code{\link{ColumnSchema}}.}
\usage{
IdObject
}
\description{
IdObject
}
\section{Fields}{

\describe{
\item{\code{id}}{of type String.}
}}

\keyword{datasets}
