% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_operator.R
\docType{data}
\name{JoinOperator}
\alias{JoinOperator}
\title{JoinOperator}
\format{\code{\link{R6Class}} object, sub classes \code{\link{JoinMeltOperator}}, \code{\link{JoinOperator1}}.}
\usage{
JoinOperator
}
\description{
JoinOperator
}
\keyword{datasets}
