% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt_step_model.R
\docType{data}
\name{MeltStepModel}
\alias{MeltStepModel}
\title{MeltStepModel}
\format{\code{\link{R6Class}} object, super class \code{\link{StepModel}}.}
\usage{
MeltStepModel
}
\description{
MeltStepModel
}
\section{Fields}{

\describe{
\item{\code{namespace}}{of type String.}

\item{\code{selectionPattern}}{of type String.}

\item{\code{factorType}}{of type String.}

\item{\code{factors}}{list of class \code{\link{Factor}}.}
}}

\keyword{datasets}
