% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator_settings.R
\docType{data}
\name{OperatorSettings}
\alias{OperatorSettings}
\title{OperatorSettings}
\format{\code{\link{R6Class}} object.}
\usage{
OperatorSettings
}
\description{
OperatorSettings
}
\section{Fields}{

\describe{
\item{\code{namespace}}{of type String.}

\item{\code{operatorRef}}{object of class \code{\link{OperatorRef}}.}
}}

\keyword{datasets}
