% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pending_state.R
\docType{data}
\name{PendingState}
\alias{PendingState}
\title{PendingState}
\format{\code{\link{R6Class}} object, super class \code{\link{State}}.}
\usage{
PendingState
}
\description{
PendingState
}
\keyword{datasets}
