% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistent_object.R
\docType{data}
\name{PersistentObject}
\alias{PersistentObject}
\title{PersistentObject}
\format{\code{\link{R6Class}} object, super class \code{\link{IdObject}}, sub classes \code{\link{ComputationTask}}, \code{\link{CSVTask}}, \code{\link{CubeQueryTask}}, \code{\link{Team}}, \code{\link{RSourceOperator}}, \code{\link{SourceOperator}}, \code{\link{ExternalOperator}}, \code{\link{CubeQueryTableSchema}}, \code{\link{ComputedTableSchema}}, \code{\link{Operator}}, \code{\link{TableSchema}}, \code{\link{FileDocument}}, \code{\link{Workflow}}, \code{\link{User}}, \code{\link{ProjectDocument}}, \code{\link{Project}}, \code{\link{Task}}, \code{\link{Document}}, \code{\link{UserSecret}}.}
\usage{
PersistentObject
}
\description{
PersistentObject
}
\section{Fields}{

\describe{
\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{isDeleted}}{of type bool.}

\item{\code{rev}}{of type String.}
}}

\keyword{datasets}
