% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistent_object.R
\docType{data}
\name{PersistentObject}
\alias{PersistentObject}
\title{PersistentObject}
\format{\code{\link{R6Class}} object, super class \code{\link{IdObject}}, sub classes \code{\link{GarbageTasks}}, \code{\link{ComputationTask}}, \code{\link{RunComputationTask}}, \code{\link{SaveComputationResultTask}}, \code{\link{ExportWorkflowTask}}, \code{\link{RunWebAppTask}}, \code{\link{CSVTask}}, \code{\link{CubeQueryTask}}, \code{\link{ExportTableTask}}, \code{\link{ImportWorkflowTask}}, \code{\link{ImportGitWorkflowTask}}, \code{\link{GlTask}}, \code{\link{ProjectTask}}, \code{\link{CreateGitOperatorTask}}, \code{\link{TaskStateEvent}}, \code{\link{TaskLogEvent}}, \code{\link{TaskProgressEvent}}, \code{\link{TaskDataEvent}}, \code{\link{TaskEvent}}, \code{\link{GenericEvent}}, \code{\link{Team}}, \code{\link{ShinyOperator}}, \code{\link{DockerOperator}}, \code{\link{ROperator}}, \code{\link{WebAppOperator}}, \code{\link{GitOperator}}, \code{\link{CubeQueryTableSchema}}, \code{\link{TableSchema}}, \code{\link{ComputedTableSchema}}, \code{\link{Schema}}, \code{\link{FileDocument}}, \code{\link{Workflow}}, \code{\link{User}}, \code{\link{Operator}}, \code{\link{WorkerEndpoint}}, \code{\link{ProjectDocument}}, \code{\link{Project}}, \code{\link{GarbageObject}}, \code{\link{Task}}, \code{\link{Event}}, \code{\link{Document}}, \code{\link{Lock}}, \code{\link{UserSecret}}.}
\usage{
PersistentObject
}
\description{
PersistentObject
}
\section{Fields}{

\describe{
\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{isDeleted}}{of type bool.}

\item{\code{rev}}{of type String.}
}}

\keyword{datasets}
