% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal.R
\docType{data}
\name{Principal}
\alias{Principal}
\title{Principal}
\format{\code{\link{R6Class}} object.}
\usage{
Principal
}
\description{
Principal
}
\section{Fields}{

\describe{
\item{\code{principalId}}{of type String.}
}}

\keyword{datasets}
