% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_service.R
\docType{data}
\name{ProjectService}
\alias{ProjectService}
\title{ProjectService}
\format{\code{\link{R6Class}} object.}
\usage{
ProjectService
}
\description{
ProjectService
}
\section{Methods}{

\describe{
   \item{\code{profiles(projectId)}}{method}
   \item{\code{resourceSummary(projectId)}}{method}
   \item{\code{explore(category,start,limit)}}{method}
   \item{\code{recentProjects(userId)}}{method}
   \item{\code{cloneProject(projectId,project)}}{method}
}
}

\keyword{datasets}
