% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relation.R
\docType{data}
\name{Relation}
\alias{Relation}
\title{Relation}
\format{\code{\link{R6Class}} object, super class \code{\link{IdObject}}, sub classes \code{\link{WhereRelation}}, \code{\link{RenameRelation}}, \code{\link{DistinctRelation}}, \code{\link{UnionRelation}}, \code{\link{SimpleRelation}}, \code{\link{InMemoryRelation}}, \code{\link{CompositeRelation}}, \code{\link{GroupByRelation}}.}
\usage{
Relation
}
\description{
Relation
}
\section{Fields}{

\describe{
\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}
}}

\keyword{datasets}
