% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_relation.R
\docType{data}
\name{SimpleRelation}
\alias{SimpleRelation}
\title{SimpleRelation}
\format{\code{\link{R6Class}} object, super class \code{\link{Relation}}.}
\usage{
SimpleRelation
}
\description{
SimpleRelation
}
\section{Fields}{

\describe{
\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}
}}

\keyword{datasets}
