% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_model.R
\docType{data}
\name{StepModel}
\alias{StepModel}
\title{StepModel}
\format{\code{\link{R6Class}} object, sub classes \code{\link{TableStepModel}}, \code{\link{Crosstab}}, \code{\link{JoinStepModel}}, \code{\link{WizardStepModel}}, \code{\link{UrlModel}}, \code{\link{MeltStepModel}}.}
\usage{
StepModel
}
\description{
StepModel
}
\keyword{datasets}
