% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_profile.R
\docType{data}
\name{StorageProfile}
\alias{StorageProfile}
\title{StorageProfile}
\format{\code{\link{R6Class}} object, super class \code{\link{Profile}}.}
\usage{
StorageProfile
}
\description{
StorageProfile
}
\section{Fields}{

\describe{
\item{\code{name}}{of type String inherited from super class \code{\link{Profile}}.}

\item{\code{size}}{of type int.}
}}

\keyword{datasets}
