% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_property.R
\docType{data}
\name{StringProperty}
\alias{StringProperty}
\title{StringProperty}
\format{\code{\link{R6Class}} object, super class \code{\link{Property}}, sub classes \code{\link{EnumeratedProperty}}.}
\usage{
StringProperty
}
\description{
StringProperty
}
\section{Fields}{

\describe{
\item{\code{name}}{of type String inherited from super class \code{\link{Property}}.}

\item{\code{description}}{of type String inherited from super class \code{\link{Property}}.}

\item{\code{defaultValue}}{of type String.}
}}

\keyword{datasets}
