% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\docType{data}
\name{Summary}
\alias{Summary}
\title{Summary}
\format{\code{\link{R6Class}} object.}
\usage{
Summary
}
\description{
Summary
}
\section{Fields}{

\describe{
\item{\code{tableSummary}}{object of class \code{\link{TableSummary}}.}

\item{\code{computedTableSummary}}{object of class \code{\link{TableSummary}}.}

\item{\code{queryTableSummary}}{object of class \code{\link{TableSummary}}.}

\item{\code{taskSummary}}{object of class \code{\link{TaskSummary}}.}
}}

\keyword{datasets}
