% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_summary.R
\docType{data}
\name{TableSummary}
\alias{TableSummary}
\title{TableSummary}
\format{\code{\link{R6Class}} object.}
\usage{
TableSummary
}
\description{
TableSummary
}
\section{Fields}{

\describe{
\item{\code{n}}{of type double.}

\item{\code{size}}{of type double.}

\item{\code{nr}}{of type double.}

\item{\code{nc}}{of type double.}
}}

\keyword{datasets}
