% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\docType{data}
\name{Task}
\alias{Task}
\title{Task}
\format{\code{\link{R6Class}} object, super class \code{\link{PersistentObject}}, sub classes \code{\link{ComputationTask}}, \code{\link{RunComputationTask}}, \code{\link{SaveComputationResultTask}}, \code{\link{ExportWorkflowTask}}, \code{\link{RunWebAppTask}}, \code{\link{CSVTask}}, \code{\link{CubeQueryTask}}, \code{\link{ExportTableTask}}, \code{\link{ImportWorkflowTask}}, \code{\link{ImportGitWorkflowTask}}, \code{\link{GlTask}}, \code{\link{ProjectTask}}, \code{\link{CreateGitOperatorTask}}.}
\usage{
Task
}
\description{
Task
}
\section{Fields}{

\describe{
\item{\code{isDeleted}}{of type bool inherited from super class \code{\link{PersistentObject}}.}

\item{\code{rev}}{of type String inherited from super class \code{\link{PersistentObject}}.}

\item{\code{id}}{of type String inherited from super class \code{\link{IdObject}}.}

\item{\code{duration}}{of type double.}

\item{\code{owner}}{of type String.}

\item{\code{taskHash}}{of type String.}

\item{\code{environment}}{list of class \code{\link{Pair}}.}

\item{\code{state}}{object of class \code{\link{State}}.}

\item{\code{createdDate}}{object of class \code{\link{Date}}.}

\item{\code{lastModifiedDate}}{object of class \code{\link{Date}}.}

\item{\code{runDate}}{object of class \code{\link{Date}}.}

\item{\code{completedDate}}{object of class \code{\link{Date}}.}

\item{\code{aclContext}}{object of class \code{\link{AclContext}}.}
}}

\keyword{datasets}
