% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker.R
\docType{data}
\name{Worker}
\alias{Worker}
\title{Worker}
\format{\code{\link{R6Class}} object.}
\usage{
Worker
}
\description{
Worker
}
\section{Fields}{

\describe{
\item{\code{uri}}{of type String.}

\item{\code{threads}}{of type int.}
}}

\keyword{datasets}
